@@ -1,7 +1,8 @@
 package com.xxmassdeveloper.mpchartexample.fragments;
 import android.graphics.Typeface;
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
+import androidx.annotation.NonNull;
+import androidx.fragment.app.Fragment;
 import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.MotionEvent;
@@ -21,48 +22,49 @@ import com.xxmassdeveloper.mpchartexample.custom.MyMarkerView;
 
 public class BarChartFrag extends SimpleFragment implements OnChartGestureListener {
 
+    @NonNull
     public static Fragment newInstance() {
         return new BarChartFrag();
     }
 
-    private BarChart mChart;
-    
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-        View v = inflater.inflate(R.layout.frag_simple_bar, container, false);
-        
-        // create a new chart object
-        mChart = new BarChart(getActivity());
-        mChart.getDescription().setEnabled(false);
-        mChart.setOnChartGestureListener(this);
-        
-        MyMarkerView mv = new MyMarkerView(getActivity(), R.layout.custom_marker_view);
-        mv.setChartView(mChart); // For bounds control
-        mChart.setMarker(mv);
+    private BarChart chart;
 
-        mChart.setDrawGridBackground(false);
-        mChart.setDrawBarShadow(false);
-        
-        Typeface tf = Typeface.createFromAsset(getActivity().getAssets(),"OpenSans-Light.ttf");
-        
-        mChart.setData(generateBarData(1, 20000, 12));
-        
-        Legend l = mChart.getLegend();
+    @Override
+    public View onCreateView(@NonNull LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
+        View v = inflater.inflate(R.layout.frag_simple_bar, container, false);
+
+        // create a new chart object
+        chart = new BarChart(getActivity());
+        chart.getDescription().setEnabled(false);
+        chart.setOnChartGestureListener(this);
+
+        MyMarkerView mv = new MyMarkerView(getActivity(), R.layout.custom_marker_view);
+        mv.setChartView(chart); // For bounds control
+        chart.setMarker(mv);
+
+        chart.setDrawGridBackground(false);
+        chart.setDrawBarShadow(false);
+
+        Typeface tf = Typeface.createFromAsset(context.getAssets(), "OpenSans-Light.ttf");
+
+        chart.setData(generateBarData(1, 20000, 12));
+
+        Legend l = chart.getLegend();
         l.setTypeface(tf);
-        
-        YAxis leftAxis = mChart.getAxisLeft();
+
+        YAxis leftAxis = chart.getAxisLeft();
         leftAxis.setTypeface(tf);
         leftAxis.setAxisMinimum(0f); // this replaces setStartAtZero(true)
 
-        mChart.getAxisRight().setEnabled(false);
-        
-        XAxis xAxis = mChart.getXAxis();
+        chart.getAxisRight().setEnabled(false);
+
+        XAxis xAxis = chart.getXAxis();
         xAxis.setEnabled(false);
-        
-        // programatically add the chart
+
+        // programmatically add the chart
         FrameLayout parent = v.findViewById(R.id.parentLayout);
-        parent.addView(mChart);
-        
+        parent.addView(chart);
+
         return v;
     }
 
@@ -74,12 +76,12 @@ public class BarChartFrag extends SimpleFragment implements OnChartGestureListen
     @Override
     public void onChartGestureEnd(MotionEvent me, ChartTouchListener.ChartGesture lastPerformedGesture) {
         Log.i("Gesture", "END");
-        mChart.highlightValues(null);
+        chart.highlightValues(null);
     }
 
     @Override
     public void onChartLongPressed(MotionEvent me) {
-        Log.i("LongPress", "Chart longpressed.");
+        Log.i("LongPress", "Chart long pressed.");
     }
 
     @Override
@@ -94,9 +96,9 @@ public class BarChartFrag extends SimpleFragment implements OnChartGestureListen
 
     @Override
     public void onChartFling(MotionEvent me1, MotionEvent me2, float velocityX, float velocityY) {
-        Log.i("Fling", "Chart flinged. VeloX: " + velocityX + ", VeloY: " + velocityY);
+        Log.i("Fling", "Chart fling. VelocityX: " + velocityX + ", VelocityY: " + velocityY);
     }
-   
+
     @Override
     public void onChartScale(MotionEvent me, float scaleX, float scaleY) {
         Log.i("Scale / Zoom", "ScaleX: " + scaleX + ", ScaleY: " + scaleY);
