@@ -2,7 +2,8 @@ package com.xxmassdeveloper.mpchartexample.fragments;
 
 import android.graphics.Typeface;
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
+import androidx.annotation.NonNull;
+import androidx.fragment.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -18,49 +19,51 @@ import com.xxmassdeveloper.mpchartexample.custom.MyMarkerView;
 
 public class ScatterChartFrag extends SimpleFragment {
 
+    @NonNull
     public static Fragment newInstance() {
         return new ScatterChartFrag();
     }
 
-    private ScatterChart mChart;
-    
-    @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-        View v = inflater.inflate(R.layout.frag_simple_scatter, container, false);
-        
-        mChart = v.findViewById(R.id.scatterChart1);
-        mChart.getDescription().setEnabled(false);
-        
-        Typeface tf = Typeface.createFromAsset(getActivity().getAssets(),"OpenSans-Light.ttf");
-        
-        MyMarkerView mv = new MyMarkerView(getActivity(), R.layout.custom_marker_view);
-        mv.setChartView(mChart); // For bounds control
-        mChart.setMarker(mv);
+    @SuppressWarnings("FieldCanBeLocal")
+    private ScatterChart chart;
 
-        mChart.setDrawGridBackground(false);
-        mChart.setData(generateScatterData(6, 10000, 200));
-        
-        XAxis xAxis = mChart.getXAxis();
+    @Override
+    public View onCreateView(@NonNull LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
+        View v = inflater.inflate(R.layout.frag_simple_scatter, container, false);
+
+        chart = v.findViewById(R.id.scatterChart1);
+        chart.getDescription().setEnabled(false);
+
+        Typeface tf = Typeface.createFromAsset(context.getAssets(), "OpenSans-Light.ttf");
+
+        MyMarkerView mv = new MyMarkerView(getActivity(), R.layout.custom_marker_view);
+        mv.setChartView(chart); // For bounds control
+        chart.setMarker(mv);
+
+        chart.setDrawGridBackground(false);
+        chart.setData(generateScatterData(6, 10000, 200));
+
+        XAxis xAxis = chart.getXAxis();
         xAxis.setEnabled(true);
         xAxis.setPosition(XAxisPosition.BOTTOM);
-        
-        YAxis leftAxis = mChart.getAxisLeft();
+
+        YAxis leftAxis = chart.getAxisLeft();
         leftAxis.setTypeface(tf);
-        
-        YAxis rightAxis = mChart.getAxisRight();
+
+        YAxis rightAxis = chart.getAxisRight();
         rightAxis.setTypeface(tf);
         rightAxis.setDrawGridLines(false);
-        
-        Legend l = mChart.getLegend();
+
+        Legend l = chart.getLegend();
         l.setWordWrapEnabled(true);
         l.setTypeface(tf);
         l.setFormSize(14f);
         l.setTextSize(9f);
-        
+
         // increase the space between legend & bottom and legend & content
-        l.setYOffset(13f);       
-        mChart.setExtraBottomOffset(16f);
-        
+        l.setYOffset(13f);
+        chart.setExtraBottomOffset(16f);
+
         return v;
     }
 }
