@@ -1,9 +1,11 @@
 package com.xxmassdeveloper.mpchartexample.fragments;
 
+import android.content.Context;
 import android.graphics.Color;
 import android.graphics.Typeface;
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
+import androidx.annotation.NonNull;
+import androidx.fragment.app.Fragment;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -28,58 +30,64 @@ import com.github.mikephil.charting.utils.FileUtils;
 
 import java.util.ArrayList;
 
+@SuppressWarnings("SameParameterValue")
 public abstract class SimpleFragment extends Fragment {
-    
+
     private Typeface tf;
-    
-    public SimpleFragment() {
-        
-    }
-    
+    protected Context context;
+
     @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-        tf = Typeface.createFromAsset(getActivity().getAssets(), "OpenSans-Regular.ttf");
+    public void onAttach(Context context) {
+        super.onAttach(context);
+        this.context = context;
+    }
+
+    public SimpleFragment() {
+
+    }
+
+    @Override
+    public View onCreateView(@NonNull LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
+        tf = Typeface.createFromAsset(context.getAssets(), "OpenSans-Regular.ttf");
         return super.onCreateView(inflater, container, savedInstanceState);
     }
 
     protected BarData generateBarData(int dataSets, float range, int count) {
-        
-        ArrayList<IBarDataSet> sets = new ArrayList<IBarDataSet>();
-        
+
+        ArrayList<IBarDataSet> sets = new ArrayList<>();
+
         for(int i = 0; i < dataSets; i++) {
-           
-            ArrayList<BarEntry> entries = new ArrayList<BarEntry>();
-            
-//            entries = FileUtils.loadEntriesFromAssets(getActivity().getAssets(), "stacked_bars.txt");
-            
-            for(int j = 0; j < count; j++) {        
+
+            ArrayList<BarEntry> entries = new ArrayList<>();
+
+            for(int j = 0; j < count; j++) {
                 entries.add(new BarEntry(j, (float) (Math.random() * range) + range / 4));
             }
-            
+
             BarDataSet ds = new BarDataSet(entries, getLabel(i));
             ds.setColors(ColorTemplate.VORDIPLOM_COLORS);
             sets.add(ds);
         }
-        
+
         BarData d = new BarData(sets);
         d.setValueTypeface(tf);
         return d;
     }
-    
+
     protected ScatterData generateScatterData(int dataSets, float range, int count) {
-        
-        ArrayList<IScatterDataSet> sets = new ArrayList<IScatterDataSet>();
-        
+
+        ArrayList<IScatterDataSet> sets = new ArrayList<>();
+
         ScatterChart.ScatterShape[] shapes = ScatterChart.ScatterShape.getAllDefaultShapes();
-        
+
         for(int i = 0; i < dataSets; i++) {
-           
-            ArrayList<Entry> entries = new ArrayList<Entry>();
-            
-            for(int j = 0; j < count; j++) {        
+
+            ArrayList<Entry> entries = new ArrayList<>();
+
+            for(int j = 0; j < count; j++) {
                 entries.add(new Entry(j, (float) (Math.random() * range) + range / 4));
             }
-            
+
             ScatterDataSet ds = new ScatterDataSet(entries, getLabel(i));
             ds.setScatterShapeSize(12f);
             ds.setScatterShape(shapes[i % shapes.length]);
@@ -87,82 +95,81 @@ public abstract class SimpleFragment extends Fragment {
             ds.setScatterShapeSize(9f);
             sets.add(ds);
         }
-        
+
         ScatterData d = new ScatterData(sets);
         d.setValueTypeface(tf);
         return d;
     }
-    
+
     /**
      * generates less data (1 DataSet, 4 values)
-     * @return
+     * @return PieData
      */
     protected PieData generatePieData() {
-        
+
         int count = 4;
-        
-        ArrayList<PieEntry> entries1 = new ArrayList<PieEntry>();
-        
+
+        ArrayList<PieEntry> entries1 = new ArrayList<>();
+
         for(int i = 0; i < count; i++) {
             entries1.add(new PieEntry((float) ((Math.random() * 60) + 40), "Quarter " + (i+1)));
         }
-        
+
         PieDataSet ds1 = new PieDataSet(entries1, "Quarterly Revenues 2015");
         ds1.setColors(ColorTemplate.VORDIPLOM_COLORS);
         ds1.setSliceSpace(2f);
         ds1.setValueTextColor(Color.WHITE);
         ds1.setValueTextSize(12f);
-        
+
         PieData d = new PieData(ds1);
         d.setValueTypeface(tf);
 
         return d;
     }
-    
+
     protected LineData generateLineData() {
-        
-        ArrayList<ILineDataSet> sets = new ArrayList<ILineDataSet>();
-        
-        LineDataSet ds1 = new LineDataSet(FileUtils.loadEntriesFromAssets(getActivity().getAssets(), "sine.txt"), "Sine function");
-        LineDataSet ds2 = new LineDataSet(FileUtils.loadEntriesFromAssets(getActivity().getAssets(), "cosine.txt"), "Cosine function");
-        
+
+        ArrayList<ILineDataSet> sets = new ArrayList<>();
+        LineDataSet ds1 = new LineDataSet(FileUtils.loadEntriesFromAssets(context.getAssets(), "sine.txt"), "Sine function");
+        LineDataSet ds2 = new LineDataSet(FileUtils.loadEntriesFromAssets(context.getAssets(), "cosine.txt"), "Cosine function");
+
         ds1.setLineWidth(2f);
         ds2.setLineWidth(2f);
-        
+
         ds1.setDrawCircles(false);
         ds2.setDrawCircles(false);
-        
+
         ds1.setColor(ColorTemplate.VORDIPLOM_COLORS[0]);
         ds2.setColor(ColorTemplate.VORDIPLOM_COLORS[1]);
-        
-        // load DataSets from textfiles in assets folders
+
+        // load DataSets from files in assets folder
         sets.add(ds1);
         sets.add(ds2);
-        
+
         LineData d = new LineData(sets);
         d.setValueTypeface(tf);
         return d;
     }
-    
+
     protected LineData getComplexity() {
-        
-        ArrayList<ILineDataSet> sets = new ArrayList<ILineDataSet>();
-        
-        LineDataSet ds1 = new LineDataSet(FileUtils.loadEntriesFromAssets(getActivity().getAssets(), "n.txt"), "O(n)");
-        LineDataSet ds2 = new LineDataSet(FileUtils.loadEntriesFromAssets(getActivity().getAssets(), "nlogn.txt"), "O(nlogn)");
-        LineDataSet ds3 = new LineDataSet(FileUtils.loadEntriesFromAssets(getActivity().getAssets(), "square.txt"), "O(n\u00B2)");
-        LineDataSet ds4 = new LineDataSet(FileUtils.loadEntriesFromAssets(getActivity().getAssets(), "three.txt"), "O(n\u00B3)");
-        
+
+        ArrayList<ILineDataSet> sets = new ArrayList<>();
+
+        LineDataSet ds1 = new LineDataSet(FileUtils.loadEntriesFromAssets(context.getAssets(), "n.txt"), "O(n)");
+        LineDataSet ds2 = new LineDataSet(FileUtils.loadEntriesFromAssets(context.getAssets(), "nlogn.txt"), "O(nlogn)");
+        LineDataSet ds3 = new LineDataSet(FileUtils.loadEntriesFromAssets(context.getAssets(), "square.txt"), "O(n\u00B2)");
+        LineDataSet ds4 = new LineDataSet(FileUtils.loadEntriesFromAssets(context.getAssets(), "three.txt"), "O(n\u00B3)");
+
         ds1.setColor(ColorTemplate.VORDIPLOM_COLORS[0]);
         ds2.setColor(ColorTemplate.VORDIPLOM_COLORS[1]);
         ds3.setColor(ColorTemplate.VORDIPLOM_COLORS[2]);
         ds4.setColor(ColorTemplate.VORDIPLOM_COLORS[3]);
-        
+
         ds1.setCircleColor(ColorTemplate.VORDIPLOM_COLORS[0]);
         ds2.setCircleColor(ColorTemplate.VORDIPLOM_COLORS[1]);
         ds3.setCircleColor(ColorTemplate.VORDIPLOM_COLORS[2]);
         ds4.setCircleColor(ColorTemplate.VORDIPLOM_COLORS[3]);
-        
+
         ds1.setLineWidth(2.5f);
         ds1.setCircleRadius(3f);
         ds2.setLineWidth(2.5f);
@@ -171,22 +178,21 @@ public abstract class SimpleFragment extends Fragment {
         ds3.setCircleRadius(3f);
         ds4.setLineWidth(2.5f);
         ds4.setCircleRadius(3f);
-        
-        
-        // load DataSets from textfiles in assets folders
-        sets.add(ds1);        
+
+
+        // load DataSets from files in assets folder
+        sets.add(ds1);
         sets.add(ds2);
         sets.add(ds3);
         sets.add(ds4);
-        
+
         LineData d = new LineData(sets);
         d.setValueTypeface(tf);
         return d;
     }
-    
-    private String[] mLabels = new String[] { "Company A", "Company B", "Company C", "Company D", "Company E", "Company F" };
-//    private String[] mXVals = new String[] { "Jan", "Feb", "Mar", "Apr", "May", "Jun", "Jul", "Aug", "Sep", "Okt", "Nov", "Dec" };
-    
+
+    private final String[] mLabels = new String[] { "Company A", "Company B", "Company C", "Company D", "Company E", "Company F" };
+
     private String getLabel(int i) {
         return mLabels[i];
     }
