@@ -1,14 +1,18 @@
 
 package com.xxmassdeveloper.mpchartexample;
 
+import android.Manifest;
+import android.content.Intent;
+import android.content.pm.PackageManager;
+import android.net.Uri;
 import android.os.Bundle;
+import androidx.core.content.ContextCompat;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.WindowManager;
 import android.widget.SeekBar;
 import android.widget.SeekBar.OnSeekBarChangeListener;
 import android.widget.TextView;
-import android.widget.Toast;
 
 import com.github.mikephil.charting.charts.BarChart;
 import com.github.mikephil.charting.components.XAxis;
@@ -25,8 +29,8 @@ import java.util.ArrayList;
 
 public class AnotherBarActivity extends DemoBase implements OnSeekBarChangeListener {
 
-    private BarChart mChart;
-    private SeekBar mSeekBarX, mSeekBarY;
+    private BarChart chart;
+    private SeekBar seekBarX, seekBarY;
     private TextView tvX, tvY;
 
     @Override
@@ -36,48 +40,89 @@ public class AnotherBarActivity extends DemoBase implements OnSeekBarChangeListe
                 WindowManager.LayoutParams.FLAG_FULLSCREEN);
         setContentView(R.layout.activity_barchart);
 
+        setTitle("AnotherBarActivity");
+
         tvX = findViewById(R.id.tvXMax);
         tvY = findViewById(R.id.tvYMax);
 
-        mSeekBarX = findViewById(R.id.seekBar1);
-        mSeekBarX.setOnSeekBarChangeListener(this);
+        seekBarX = findViewById(R.id.seekBar1);
+        seekBarX.setOnSeekBarChangeListener(this);
 
-        mSeekBarY = findViewById(R.id.seekBar2);
-        mSeekBarY.setOnSeekBarChangeListener(this);
+        seekBarY = findViewById(R.id.seekBar2);
+        seekBarY.setOnSeekBarChangeListener(this);
 
-        mChart = findViewById(R.id.chart1);
+        chart = findViewById(R.id.chart1);
 
-        mChart.getDescription().setEnabled(false);
+        chart.getDescription().setEnabled(false);
 
         // if more than 60 entries are displayed in the chart, no values will be
         // drawn
-        mChart.setMaxVisibleValueCount(60);
+        chart.setMaxVisibleValueCount(60);
 
         // scaling can now only be done on x- and y-axis separately
-        mChart.setPinchZoom(false);
+        chart.setPinchZoom(false);
 
-        mChart.setDrawBarShadow(false);
-        mChart.setDrawGridBackground(false);
+        chart.setDrawBarShadow(false);
+        chart.setDrawGridBackground(false);
 
-        XAxis xAxis = mChart.getXAxis();
+        XAxis xAxis = chart.getXAxis();
         xAxis.setPosition(XAxisPosition.BOTTOM);
         xAxis.setDrawGridLines(false);
-        
-        mChart.getAxisLeft().setDrawGridLines(false);
+
+        chart.getAxisLeft().setDrawGridLines(false);
 
         // setting data
-        mSeekBarX.setProgress(10);
-        mSeekBarY.setProgress(100);
+        seekBarX.setProgress(10);
+        seekBarY.setProgress(100);
 
         // add a nice and smooth animation
-        mChart.animateY(2500);
-        
-        mChart.getLegend().setEnabled(false);
+        chart.animateY(1500);
+
+        chart.getLegend().setEnabled(false);
+    }
+
+    @Override
+    public void onProgressChanged(SeekBar seekBar, int progress, boolean fromUser) {
+
+        tvX.setText(String.valueOf(seekBarX.getProgress()));
+        tvY.setText(String.valueOf(seekBarY.getProgress()));
+
+        ArrayList<BarEntry> values = new ArrayList<>();
+
+        for (int i = 0; i < seekBarX.getProgress(); i++) {
+            float multi = (seekBarY.getProgress() + 1);
+            float val = (float) (Math.random() * multi) + multi / 3;
+            values.add(new BarEntry(i, val));
+        }
+
+        BarDataSet set1;
+
+        if (chart.getData() != null &&
+                chart.getData().getDataSetCount() > 0) {
+            set1 = (BarDataSet) chart.getData().getDataSetByIndex(0);
+            set1.setValues(values);
+            chart.getData().notifyDataChanged();
+            chart.notifyDataSetChanged();
+        } else {
+            set1 = new BarDataSet(values, "Data Set");
+            set1.setColors(ColorTemplate.VORDIPLOM_COLORS);
+            set1.setDrawValues(false);
+
+            ArrayList<IBarDataSet> dataSets = new ArrayList<>();
+            dataSets.add(set1);
+
+            BarData data = new BarData(dataSets);
+            chart.setData(data);
+            chart.setFitBars(true);
+        }
+
+        chart.invalidate();
     }
 
     @Override
     public boolean onCreateOptionsMenu(Menu menu) {
         getMenuInflater().inflate(R.menu.bar, menu);
+        menu.removeItem(R.id.actionToggleIcons);
         return true;
     }
 
@@ -85,63 +130,71 @@ public class AnotherBarActivity extends DemoBase implements OnSeekBarChangeListe
     public boolean onOptionsItemSelected(MenuItem item) {
 
         switch (item.getItemId()) {
-            case R.id.actionToggleValues: {
-
-                for (IDataSet set : mChart.getData().getDataSets())
-                    set.setDrawValues(!set.isDrawValuesEnabled());
-
-                mChart.invalidate();
+            case R.id.viewGithub: {
+                Intent i = new Intent(Intent.ACTION_VIEW);
+                i.setData(Uri.parse("https://github.com/PhilJay/MPAndroidChart/blob/master/MPChartExample/src/com/xxmassdeveloper/mpchartexample/AnotherBarActivity.java"));
+                startActivity(i);
                 break;
             }
+            case R.id.actionToggleValues: {
+
+                for (IDataSet set : chart.getData().getDataSets())
+                    set.setDrawValues(!set.isDrawValuesEnabled());
+
+                chart.invalidate();
+                break;
+            }
+            /*
+            case R.id.actionToggleIcons: { break; }
+             */
             case R.id.actionToggleHighlight: {
 
-                if(mChart.getData() != null) {
-                    mChart.getData().setHighlightEnabled(!mChart.getData().isHighlightEnabled());
-                    mChart.invalidate();
+                if(chart.getData() != null) {
+                    chart.getData().setHighlightEnabled(!chart.getData().isHighlightEnabled());
+                    chart.invalidate();
                 }
                 break;
             }
             case R.id.actionTogglePinch: {
-                if (mChart.isPinchZoomEnabled())
-                    mChart.setPinchZoom(false);
+                if (chart.isPinchZoomEnabled())
+                    chart.setPinchZoom(false);
                 else
-                    mChart.setPinchZoom(true);
+                    chart.setPinchZoom(true);
 
-                mChart.invalidate();
+                chart.invalidate();
                 break;
             }
             case R.id.actionToggleAutoScaleMinMax: {
-                mChart.setAutoScaleMinMaxEnabled(!mChart.isAutoScaleMinMaxEnabled());
-                mChart.notifyDataSetChanged();
+                chart.setAutoScaleMinMaxEnabled(!chart.isAutoScaleMinMaxEnabled());
+                chart.notifyDataSetChanged();
                 break;
             }
             case R.id.actionToggleBarBorders: {
-                for (IBarDataSet set : mChart.getData().getDataSets())
+                for (IBarDataSet set : chart.getData().getDataSets())
                     ((BarDataSet)set).setBarBorderWidth(set.getBarBorderWidth() == 1.f ? 0.f : 1.f);
 
-                mChart.invalidate();
+                chart.invalidate();
                 break;
             }
             case R.id.animateX: {
-                mChart.animateX(3000);
+                chart.animateX(2000);
                 break;
             }
             case R.id.animateY: {
-                mChart.animateY(3000);
+                chart.animateY(2000);
                 break;
             }
             case R.id.animateXY: {
 
-                mChart.animateXY(3000, 3000);
+                chart.animateXY(2000, 2000);
                 break;
             }
             case R.id.actionSave: {
-                if (mChart.saveToGallery("title" + System.currentTimeMillis(), 50)) {
-                    Toast.makeText(getApplicationContext(), "Saving SUCCESSFUL!",
-                            Toast.LENGTH_SHORT).show();
-                } else
-                    Toast.makeText(getApplicationContext(), "Saving FAILED!", Toast.LENGTH_SHORT)
-                            .show();
+                if (ContextCompat.checkSelfPermission(this, Manifest.permission.WRITE_EXTERNAL_STORAGE) == PackageManager.PERMISSION_GRANTED) {
+                    saveToGallery();
+                } else {
+                    requestStoragePermission(chart);
+                }
                 break;
             }
         }
@@ -149,52 +202,13 @@ public class AnotherBarActivity extends DemoBase implements OnSeekBarChangeListe
     }
 
     @Override
-    public void onProgressChanged(SeekBar seekBar, int progress, boolean fromUser) {
-
-        tvX.setText("" + (mSeekBarX.getProgress() + 1));
-        tvY.setText("" + (mSeekBarY.getProgress()));
-
-        ArrayList<BarEntry> yVals1 = new ArrayList<BarEntry>();
-
-        for (int i = 0; i < mSeekBarX.getProgress() + 1; i++) {
-            float mult = (mSeekBarY.getProgress() + 1);
-            float val = (float) (Math.random() * mult) + mult / 3;
-            yVals1.add(new BarEntry(i, val));
-        }
-
-        BarDataSet set1;
-
-        if (mChart.getData() != null &&
-                mChart.getData().getDataSetCount() > 0) {
-            set1 = (BarDataSet)mChart.getData().getDataSetByIndex(0);
-            set1.setValues(yVals1);
-            mChart.getData().notifyDataChanged();
-            mChart.notifyDataSetChanged();
-        } else {
-            set1 = new BarDataSet(yVals1, "Data Set");
-            set1.setColors(ColorTemplate.VORDIPLOM_COLORS);
-            set1.setDrawValues(false);
-
-            ArrayList<IBarDataSet> dataSets = new ArrayList<IBarDataSet>();
-            dataSets.add(set1);
-
-            BarData data = new BarData(dataSets);
-            mChart.setData(data);
-            mChart.setFitBars(true);
-        }
-
-        mChart.invalidate();
+    public void saveToGallery() {
+        saveToGallery(chart, "AnotherBarActivity");
     }
 
     @Override
-    public void onStartTrackingTouch(SeekBar seekBar) {
-        // TODO Auto-generated method stub
-
-    }
+    public void onStartTrackingTouch(SeekBar seekBar) {}
 
     @Override
-    public void onStopTrackingTouch(SeekBar seekBar) {
-        // TODO Auto-generated method stub
-
-    }
+    public void onStopTrackingTouch(SeekBar seekBar) {}
 }
