@@ -1,8 +1,13 @@
 
 package com.xxmassdeveloper.mpchartexample;
 
+import android.Manifest;
+import android.content.Intent;
+import android.content.pm.PackageManager;
 import android.graphics.Color;
+import android.net.Uri;
 import android.os.Bundle;
+import androidx.core.content.ContextCompat;
 import android.util.Log;
 import android.view.Menu;
 import android.view.MenuItem;
@@ -29,12 +34,13 @@ import com.xxmassdeveloper.mpchartexample.custom.MyMarkerView;
 import com.xxmassdeveloper.mpchartexample.notimportant.DemoBase;
 
 import java.util.ArrayList;
+import java.util.Locale;
 
 public class BarChartActivityMultiDataset extends DemoBase implements OnSeekBarChangeListener,
         OnChartValueSelectedListener {
 
-    private BarChart mChart;
-    private SeekBar mSeekBarX, mSeekBarY;
+    private BarChart chart;
+    private SeekBar seekBarX, seekBarY;
     private TextView tvX, tvY;
 
     @Override
@@ -44,51 +50,53 @@ public class BarChartActivityMultiDataset extends DemoBase implements OnSeekBarC
                 WindowManager.LayoutParams.FLAG_FULLSCREEN);
         setContentView(R.layout.activity_barchart);
 
+        setTitle("BarChartActivityMultiDataset");
+
         tvX = findViewById(R.id.tvXMax);
         tvX.setTextSize(10);
         tvY = findViewById(R.id.tvYMax);
 
-        mSeekBarX = findViewById(R.id.seekBar1);
-        mSeekBarX.setOnSeekBarChangeListener(this);
+        seekBarX = findViewById(R.id.seekBar1);
+        seekBarX.setOnSeekBarChangeListener(this);
 
-        mSeekBarY = findViewById(R.id.seekBar2);
-        mSeekBarY.setOnSeekBarChangeListener(this);
+        seekBarY = findViewById(R.id.seekBar2);
+        seekBarY.setOnSeekBarChangeListener(this);
 
-        mChart = findViewById(R.id.chart1);
-        mChart.setOnChartValueSelectedListener(this);
-        mChart.getDescription().setEnabled(false);
+        chart = findViewById(R.id.chart1);
+        chart.setOnChartValueSelectedListener(this);
+        chart.getDescription().setEnabled(false);
 
-//        mChart.setDrawBorders(true);
+//        chart.setDrawBorders(true);
 
         // scaling can now only be done on x- and y-axis separately
-        mChart.setPinchZoom(false);
+        chart.setPinchZoom(false);
 
-        mChart.setDrawBarShadow(false);
+        chart.setDrawBarShadow(false);
 
-        mChart.setDrawGridBackground(false);
+        chart.setDrawGridBackground(false);
 
         // create a custom MarkerView (extend MarkerView) and specify the layout
         // to use for it
         MyMarkerView mv = new MyMarkerView(this, R.layout.custom_marker_view);
-        mv.setChartView(mChart); // For bounds control
-        mChart.setMarker(mv); // Set the marker to the chart
+        mv.setChartView(chart); // For bounds control
+        chart.setMarker(mv); // Set the marker to the chart
 
-        mSeekBarX.setProgress(10);
-        mSeekBarY.setProgress(100);
+        seekBarX.setProgress(10);
+        seekBarY.setProgress(100);
 
-        Legend l = mChart.getLegend();
+        Legend l = chart.getLegend();
         l.setVerticalAlignment(Legend.LegendVerticalAlignment.TOP);
         l.setHorizontalAlignment(Legend.LegendHorizontalAlignment.RIGHT);
         l.setOrientation(Legend.LegendOrientation.VERTICAL);
         l.setDrawInside(true);
-        l.setTypeface(mTfLight);
+        l.setTypeface(tfLight);
         l.setYOffset(0f);
         l.setXOffset(10f);
         l.setYEntrySpace(0f);
         l.setTextSize(8f);
 
-        XAxis xAxis = mChart.getXAxis();
-        xAxis.setTypeface(mTfLight);
+        XAxis xAxis = chart.getXAxis();
+        xAxis.setTypeface(tfLight);
         xAxis.setGranularity(1f);
         xAxis.setCenterAxisLabels(true);
         xAxis.setValueFormatter(new IAxisValueFormatter() {
@@ -98,14 +106,88 @@ public class BarChartActivityMultiDataset extends DemoBase implements OnSeekBarC
             }
         });
 
-        YAxis leftAxis = mChart.getAxisLeft();
-        leftAxis.setTypeface(mTfLight);
+        YAxis leftAxis = chart.getAxisLeft();
+        leftAxis.setTypeface(tfLight);
         leftAxis.setValueFormatter(new LargeValueFormatter());
         leftAxis.setDrawGridLines(false);
         leftAxis.setSpaceTop(35f);
         leftAxis.setAxisMinimum(0f); // this replaces setStartAtZero(true)
 
-        mChart.getAxisRight().setEnabled(false);
+        chart.getAxisRight().setEnabled(false);
+    }
+
+    @Override
+    public void onProgressChanged(SeekBar seekBar, int progress, boolean fromUser) {
+
+        float groupSpace = 0.08f;
+        float barSpace = 0.03f; // x4 DataSet
+        float barWidth = 0.2f; // x4 DataSet
+        // (0.2 + 0.03) * 4 + 0.08 = 1.00 -> interval per "group"
+
+        int groupCount = seekBarX.getProgress() + 1;
+        int startYear = 1980;
+        int endYear = startYear + groupCount;
+
+        tvX.setText(String.format(Locale.ENGLISH, "%d-%d", startYear, endYear));
+        tvY.setText(String.valueOf(seekBarY.getProgress()));
+
+        ArrayList<BarEntry> values1 = new ArrayList<>();
+        ArrayList<BarEntry> values2 = new ArrayList<>();
+        ArrayList<BarEntry> values3 = new ArrayList<>();
+        ArrayList<BarEntry> values4 = new ArrayList<>();
+
+        float randomMultiplier = seekBarY.getProgress() * 100000f;
+
+        for (int i = startYear; i < endYear; i++) {
+            values1.add(new BarEntry(i, (float) (Math.random() * randomMultiplier)));
+            values2.add(new BarEntry(i, (float) (Math.random() * randomMultiplier)));
+            values3.add(new BarEntry(i, (float) (Math.random() * randomMultiplier)));
+            values4.add(new BarEntry(i, (float) (Math.random() * randomMultiplier)));
+        }
+
+        BarDataSet set1, set2, set3, set4;
+
+        if (chart.getData() != null && chart.getData().getDataSetCount() > 0) {
+
+            set1 = (BarDataSet) chart.getData().getDataSetByIndex(0);
+            set2 = (BarDataSet) chart.getData().getDataSetByIndex(1);
+            set3 = (BarDataSet) chart.getData().getDataSetByIndex(2);
+            set4 = (BarDataSet) chart.getData().getDataSetByIndex(3);
+            set1.setValues(values1);
+            set2.setValues(values2);
+            set3.setValues(values3);
+            set4.setValues(values4);
+            chart.getData().notifyDataChanged();
+            chart.notifyDataSetChanged();
+
+        } else {
+            // create 4 DataSets
+            set1 = new BarDataSet(values1, "Company A");
+            set1.setColor(Color.rgb(104, 241, 175));
+            set2 = new BarDataSet(values2, "Company B");
+            set2.setColor(Color.rgb(164, 228, 251));
+            set3 = new BarDataSet(values3, "Company C");
+            set3.setColor(Color.rgb(242, 247, 158));
+            set4 = new BarDataSet(values4, "Company D");
+            set4.setColor(Color.rgb(255, 102, 0));
+
+            BarData data = new BarData(set1, set2, set3, set4);
+            data.setValueFormatter(new LargeValueFormatter());
+            data.setValueTypeface(tfLight);
+
+            chart.setData(data);
+        }
+
+        // specify the width each bar should have
+        chart.getBarData().setBarWidth(barWidth);
+
+        // restrict the x-axis range
+        chart.getXAxis().setAxisMinimum(startYear);
+
+        // barData.getGroupWith(...) is a helper that calculates the width each group needs based on the provided parameters
+        chart.getXAxis().setAxisMaximum(startYear + chart.getBarData().getGroupWidth(groupSpace, barSpace) * groupCount);
+        chart.groupBars(startYear, groupSpace, barSpace);
+        chart.invalidate();
     }
 
     @Override
@@ -118,56 +200,65 @@ public class BarChartActivityMultiDataset extends DemoBase implements OnSeekBarC
     public boolean onOptionsItemSelected(MenuItem item) {
 
         switch (item.getItemId()) {
+            case R.id.viewGithub: {
+                Intent i = new Intent(Intent.ACTION_VIEW);
+                i.setData(Uri.parse("https://github.com/PhilJay/MPAndroidChart/blob/master/MPChartExample/src/com/xxmassdeveloper/mpchartexample/BarChartActivityMultiDataset.java"));
+                startActivity(i);
+                break;
+            }
             case R.id.actionToggleValues: {
-                for (IBarDataSet set : mChart.getData().getDataSets())
+                for (IBarDataSet set : chart.getData().getDataSets())
                     set.setDrawValues(!set.isDrawValuesEnabled());
 
-                mChart.invalidate();
+                chart.invalidate();
                 break;
             }
             case R.id.actionTogglePinch: {
-                if (mChart.isPinchZoomEnabled())
-                    mChart.setPinchZoom(false);
+                if (chart.isPinchZoomEnabled())
+                    chart.setPinchZoom(false);
                 else
-                    mChart.setPinchZoom(true);
+                    chart.setPinchZoom(true);
 
-                mChart.invalidate();
+                chart.invalidate();
                 break;
             }
             case R.id.actionToggleAutoScaleMinMax: {
-                mChart.setAutoScaleMinMaxEnabled(!mChart.isAutoScaleMinMaxEnabled());
-                mChart.notifyDataSetChanged();
+                chart.setAutoScaleMinMaxEnabled(!chart.isAutoScaleMinMaxEnabled());
+                chart.notifyDataSetChanged();
                 break;
             }
             case R.id.actionToggleBarBorders: {
-                for (IBarDataSet set : mChart.getData().getDataSets())
+                for (IBarDataSet set : chart.getData().getDataSets())
                     ((BarDataSet) set).setBarBorderWidth(set.getBarBorderWidth() == 1.f ? 0.f : 1.f);
 
-                mChart.invalidate();
+                chart.invalidate();
                 break;
             }
             case R.id.actionToggleHighlight: {
-                if (mChart.getData() != null) {
-                    mChart.getData().setHighlightEnabled(!mChart.getData().isHighlightEnabled());
-                    mChart.invalidate();
+                if (chart.getData() != null) {
+                    chart.getData().setHighlightEnabled(!chart.getData().isHighlightEnabled());
+                    chart.invalidate();
                 }
                 break;
             }
             case R.id.actionSave: {
-                // mChart.saveToGallery("title"+System.currentTimeMillis());
-                mChart.saveToPath("title" + System.currentTimeMillis(), "");
+                if (ContextCompat.checkSelfPermission(this, Manifest.permission.WRITE_EXTERNAL_STORAGE) == PackageManager.PERMISSION_GRANTED) {
+                    saveToGallery();
+                } else {
+                    requestStoragePermission(chart);
+                }
                 break;
             }
             case R.id.animateX: {
-                mChart.animateX(3000);
+                chart.animateX(2000);
                 break;
             }
             case R.id.animateY: {
-                mChart.animateY(3000);
+                chart.animateY(2000);
                 break;
             }
             case R.id.animateXY: {
-                mChart.animateXY(3000, 3000);
+                chart.animateXY(2000, 2000);
                 break;
             }
         }
@@ -175,88 +266,15 @@ public class BarChartActivityMultiDataset extends DemoBase implements OnSeekBarC
     }
 
     @Override
-    public void onProgressChanged(SeekBar seekBar, int progress, boolean fromUser) {
-
-        float groupSpace = 0.08f;
-        float barSpace = 0.03f; // x4 DataSet
-        float barWidth = 0.2f; // x4 DataSet
-        // (0.2 + 0.03) * 4 + 0.08 = 1.00 -> interval per "group"
-
-        int groupCount = mSeekBarX.getProgress() + 1;
-        int startYear = 1980;
-        int endYear = startYear + groupCount;
-
-        tvX.setText(startYear + "-" + endYear);
-        tvY.setText("" + (mSeekBarY.getProgress()));
-
-        ArrayList<BarEntry> yVals1 = new ArrayList<BarEntry>();
-        ArrayList<BarEntry> yVals2 = new ArrayList<BarEntry>();
-        ArrayList<BarEntry> yVals3 = new ArrayList<BarEntry>();
-        ArrayList<BarEntry> yVals4 = new ArrayList<BarEntry>();
-
-        float randomMultiplier = mSeekBarY.getProgress() * 100000f;
-
-        for (int i = startYear; i < endYear; i++) {
-            yVals1.add(new BarEntry(i, (float) (Math.random() * randomMultiplier)));
-            yVals2.add(new BarEntry(i, (float) (Math.random() * randomMultiplier)));
-            yVals3.add(new BarEntry(i, (float) (Math.random() * randomMultiplier)));
-            yVals4.add(new BarEntry(i, (float) (Math.random() * randomMultiplier)));
-        }
-
-        BarDataSet set1, set2, set3, set4;
-
-        if (mChart.getData() != null && mChart.getData().getDataSetCount() > 0) {
-
-            set1 = (BarDataSet) mChart.getData().getDataSetByIndex(0);
-            set2 = (BarDataSet) mChart.getData().getDataSetByIndex(1);
-            set3 = (BarDataSet) mChart.getData().getDataSetByIndex(2);
-            set4 = (BarDataSet) mChart.getData().getDataSetByIndex(3);
-            set1.setValues(yVals1);
-            set2.setValues(yVals2);
-            set3.setValues(yVals3);
-            set4.setValues(yVals4);
-            mChart.getData().notifyDataChanged();
-            mChart.notifyDataSetChanged();
-
-        } else {
-            // create 4 DataSets
-            set1 = new BarDataSet(yVals1, "Company A");
-            set1.setColor(Color.rgb(104, 241, 175));
-            set2 = new BarDataSet(yVals2, "Company B");
-            set2.setColor(Color.rgb(164, 228, 251));
-            set3 = new BarDataSet(yVals3, "Company C");
-            set3.setColor(Color.rgb(242, 247, 158));
-            set4 = new BarDataSet(yVals4, "Company D");
-            set4.setColor(Color.rgb(255, 102, 0));
-
-            BarData data = new BarData(set1, set2, set3, set4);
-            data.setValueFormatter(new LargeValueFormatter());
-            data.setValueTypeface(mTfLight);
-
-            mChart.setData(data);
-        }
-
-        // specify the width each bar should have
-        mChart.getBarData().setBarWidth(barWidth);
-
-        // restrict the x-axis range
-        mChart.getXAxis().setAxisMinimum(startYear);
-
-        // barData.getGroupWith(...) is a helper that calculates the width each group needs based on the provided parameters
-        mChart.getXAxis().setAxisMaximum(startYear + mChart.getBarData().getGroupWidth(groupSpace, barSpace) * groupCount);
-        mChart.groupBars(startYear, groupSpace, barSpace);
-        mChart.invalidate();
+    public void saveToGallery() {
+        saveToGallery(chart, "BarChartActivityMultiDataset");
     }
 
     @Override
-    public void onStartTrackingTouch(SeekBar seekBar) {
-        // TODO Auto-generated method stub
-    }
+    public void onStartTrackingTouch(SeekBar seekBar) {}
 
     @Override
-    public void onStopTrackingTouch(SeekBar seekBar) {
-        // TODO Auto-generated method stub
-    }
+    public void onStopTrackingTouch(SeekBar seekBar) {}
 
     @Override
     public void onValueSelected(Entry e, Highlight h) {
