@@ -1,20 +1,23 @@
 
 package com.xxmassdeveloper.mpchartexample;
 
+import android.Manifest;
+import android.content.Intent;
+import android.content.pm.PackageManager;
 import android.graphics.Color;
+import android.net.Uri;
 import android.os.Bundle;
+import androidx.core.content.ContextCompat;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.WindowManager;
 import android.widget.SeekBar;
 import android.widget.SeekBar.OnSeekBarChangeListener;
 import android.widget.TextView;
-import android.widget.Toast;
 
 import com.github.mikephil.charting.charts.BarChart;
 import com.github.mikephil.charting.components.Legend;
 import com.github.mikephil.charting.components.Legend.LegendForm;
-import com.github.mikephil.charting.components.Legend.LegendPosition;
 import com.github.mikephil.charting.components.XAxis;
 import com.github.mikephil.charting.components.YAxis;
 import com.github.mikephil.charting.data.BarData;
@@ -29,11 +32,11 @@ import java.util.List;
 
 public class BarChartActivitySinus extends DemoBase implements OnSeekBarChangeListener {
 
-    protected BarChart mChart;
-    private SeekBar mSeekBarX;
+    private BarChart chart;
+    private SeekBar seekBarX;
     private TextView tvX;
 
-    private List<BarEntry> mSinusData;
+    private List<BarEntry> data;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -42,57 +45,59 @@ public class BarChartActivitySinus extends DemoBase implements OnSeekBarChangeLi
                 WindowManager.LayoutParams.FLAG_FULLSCREEN);
         setContentView(R.layout.activity_barchart_sinus);
 
-        mSinusData = FileUtils.loadBarEntriesFromAssets(getAssets(), "othersine.txt");
+        setTitle("BarChartActivitySinus");
+
+        data = FileUtils.loadBarEntriesFromAssets(getAssets(), "othersine.txt");
 
         tvX = findViewById(R.id.tvValueCount);
 
-        mSeekBarX = findViewById(R.id.seekbarValues);
+        seekBarX = findViewById(R.id.seekbarValues);
 
-        mChart = findViewById(R.id.chart1);
+        chart = findViewById(R.id.chart1);
 
-        mChart.setDrawBarShadow(false);
-        mChart.setDrawValueAboveBar(true);
+        chart.setDrawBarShadow(false);
+        chart.setDrawValueAboveBar(true);
 
-        mChart.getDescription().setEnabled(false);
+        chart.getDescription().setEnabled(false);
 
         // if more than 60 entries are displayed in the chart, no values will be
         // drawn
-        mChart.setMaxVisibleValueCount(60);
+        chart.setMaxVisibleValueCount(60);
 
         // scaling can now only be done on x- and y-axis separately
-        mChart.setPinchZoom(false);
+        chart.setPinchZoom(false);
 
         // draw shadows for each bar that show the maximum value
-        // mChart.setDrawBarShadow(true);
+        // chart.setDrawBarShadow(true);
 
-        // mChart.setDrawXLabels(false);
+        // chart.setDrawXLabels(false);
 
-        mChart.setDrawGridBackground(false);
-        // mChart.setDrawYLabels(false);
+        chart.setDrawGridBackground(false);
+        // chart.setDrawYLabels(false);
 
-        XAxis xAxis = mChart.getXAxis();
+        XAxis xAxis = chart.getXAxis();
         xAxis.setEnabled(false);
 
-        YAxis leftAxis = mChart.getAxisLeft();
-        leftAxis.setTypeface(mTfLight);
+        YAxis leftAxis = chart.getAxisLeft();
+        leftAxis.setTypeface(tfLight);
         leftAxis.setLabelCount(6, false);
         leftAxis.setAxisMinimum(-2.5f);
         leftAxis.setAxisMaximum(2.5f);
         leftAxis.setGranularityEnabled(true);
         leftAxis.setGranularity(0.1f);
 
-        YAxis rightAxis = mChart.getAxisRight();
+        YAxis rightAxis = chart.getAxisRight();
         rightAxis.setDrawGridLines(false);
-        rightAxis.setTypeface(mTfLight);
+        rightAxis.setTypeface(tfLight);
         rightAxis.setLabelCount(6, false);
         rightAxis.setAxisMinimum(-2.5f);
         rightAxis.setAxisMaximum(2.5f);
         rightAxis.setGranularity(0.1f);
 
-        mSeekBarX.setOnSeekBarChangeListener(this);
-        mSeekBarX.setProgress(150); // set data
+        seekBarX.setOnSeekBarChangeListener(this);
+        seekBarX.setProgress(150); // set data
 
-        Legend l = mChart.getLegend();
+        Legend l = chart.getLegend();
         l.setVerticalAlignment(Legend.LegendVerticalAlignment.BOTTOM);
         l.setHorizontalAlignment(Legend.LegendHorizontalAlignment.LEFT);
         l.setOrientation(Legend.LegendOrientation.HORIZONTAL);
@@ -102,7 +107,37 @@ public class BarChartActivitySinus extends DemoBase implements OnSeekBarChangeLi
         l.setTextSize(11f);
         l.setXEntrySpace(4f);
 
-        mChart.animateXY(2000, 2000);
+        chart.animateXY(1500, 1500);
+    }
+
+    private void setData(int count) {
+
+        ArrayList<BarEntry> entries = new ArrayList<>();
+
+        for (int i = 0; i < count; i++) {
+            entries.add(data.get(i));
+        }
+
+        BarDataSet set;
+
+        if (chart.getData() != null &&
+                chart.getData().getDataSetCount() > 0) {
+            set = (BarDataSet) chart.getData().getDataSetByIndex(0);
+            set.setValues(entries);
+            chart.getData().notifyDataChanged();
+            chart.notifyDataSetChanged();
+        } else {
+            set = new BarDataSet(entries, "Sinus Function");
+            set.setColor(Color.rgb(240, 120, 124));
+        }
+
+        BarData data = new BarData(set);
+        data.setValueTextSize(10f);
+        data.setValueTypeface(tfLight);
+        data.setDrawValues(false);
+        data.setBarWidth(0.8f);
+
+        chart.setData(data);
     }
 
     @Override
@@ -115,61 +150,66 @@ public class BarChartActivitySinus extends DemoBase implements OnSeekBarChangeLi
     public boolean onOptionsItemSelected(MenuItem item) {
 
         switch (item.getItemId()) {
+            case R.id.viewGithub: {
+                Intent i = new Intent(Intent.ACTION_VIEW);
+                i.setData(Uri.parse("https://github.com/PhilJay/MPAndroidChart/blob/master/MPChartExample/src/com/xxmassdeveloper/mpchartexample/BarChartActivitySinus.java"));
+                startActivity(i);
+                break;
+            }
             case R.id.actionToggleValues: {
-                for (IBarDataSet set : mChart.getData().getDataSets())
+                for (IBarDataSet set : chart.getData().getDataSets())
                     set.setDrawValues(!set.isDrawValuesEnabled());
 
-                mChart.invalidate();
+                chart.invalidate();
                 break;
             }
             case R.id.actionToggleHighlight: {
-                if (mChart.getData() != null) {
-                    mChart.getData().setHighlightEnabled(!mChart.getData().isHighlightEnabled());
-                    mChart.invalidate();
+                if (chart.getData() != null) {
+                    chart.getData().setHighlightEnabled(!chart.getData().isHighlightEnabled());
+                    chart.invalidate();
                 }
                 break;
             }
             case R.id.actionTogglePinch: {
-                if (mChart.isPinchZoomEnabled())
-                    mChart.setPinchZoom(false);
+                if (chart.isPinchZoomEnabled())
+                    chart.setPinchZoom(false);
                 else
-                    mChart.setPinchZoom(true);
+                    chart.setPinchZoom(true);
 
-                mChart.invalidate();
+                chart.invalidate();
                 break;
             }
             case R.id.actionToggleAutoScaleMinMax: {
-                mChart.setAutoScaleMinMaxEnabled(!mChart.isAutoScaleMinMaxEnabled());
-                mChart.notifyDataSetChanged();
+                chart.setAutoScaleMinMaxEnabled(!chart.isAutoScaleMinMaxEnabled());
+                chart.notifyDataSetChanged();
                 break;
             }
             case R.id.actionToggleBarBorders: {
-                for (IBarDataSet set : mChart.getData().getDataSets())
+                for (IBarDataSet set : chart.getData().getDataSets())
                     ((BarDataSet) set).setBarBorderWidth(set.getBarBorderWidth() == 1.f ? 0.f : 1.f);
 
-                mChart.invalidate();
+                chart.invalidate();
                 break;
             }
             case R.id.animateX: {
-                mChart.animateX(1500);
+                chart.animateX(2000);
                 break;
             }
             case R.id.animateY: {
-                mChart.animateY(1500);
+                chart.animateY(2000);
                 break;
             }
             case R.id.animateXY: {
 
-                mChart.animateXY(2000, 2000);
+                chart.animateXY(2000, 2000);
                 break;
             }
             case R.id.actionSave: {
-                if (mChart.saveToGallery("title" + System.currentTimeMillis(), 50)) {
-                    Toast.makeText(getApplicationContext(), "Saving SUCCESSFUL!",
-                            Toast.LENGTH_SHORT).show();
-                } else
-                    Toast.makeText(getApplicationContext(), "Saving FAILED!", Toast.LENGTH_SHORT)
-                            .show();
+                if (ContextCompat.checkSelfPermission(this, Manifest.permission.WRITE_EXTERNAL_STORAGE) == PackageManager.PERMISSION_GRANTED) {
+                    saveToGallery();
+                } else {
+                    requestStoragePermission(chart);
+                }
                 break;
             }
         }
@@ -179,51 +219,21 @@ public class BarChartActivitySinus extends DemoBase implements OnSeekBarChangeLi
     @Override
     public void onProgressChanged(SeekBar seekBar, int progress, boolean fromUser) {
 
-        tvX.setText("" + (mSeekBarX.getProgress()));
+        tvX.setText(String.valueOf(seekBarX.getProgress()));
 
-        setData(mSeekBarX.getProgress());
-        mChart.invalidate();
+        setData(seekBarX.getProgress());
+        chart.invalidate();
     }
 
     @Override
-    public void onStartTrackingTouch(SeekBar seekBar) {
-        // TODO Auto-generated method stub
-
+    public void saveToGallery() {
+        saveToGallery(chart, "BarChartActivitySinus");
     }
 
     @Override
-    public void onStopTrackingTouch(SeekBar seekBar) {
-        // TODO Auto-generated method stub
+    public void onStartTrackingTouch(SeekBar seekBar) {}
 
-    }
+    @Override
+    public void onStopTrackingTouch(SeekBar seekBar) {}
 
-    private void setData(int count) {
-
-        ArrayList<BarEntry> entries = new ArrayList<BarEntry>();
-
-        for (int i = 0; i < count; i++) {
-            entries.add(mSinusData.get(i));
-        }
-
-        BarDataSet set;
-
-        if (mChart.getData() != null &&
-                mChart.getData().getDataSetCount() > 0) {
-            set = (BarDataSet) mChart.getData().getDataSetByIndex(0);
-            set.setValues(entries);
-            mChart.getData().notifyDataChanged();
-            mChart.notifyDataSetChanged();
-        } else {
-            set = new BarDataSet(entries, "Sinus Function");
-            set.setColor(Color.rgb(240, 120, 124));
-        }
-
-        BarData data = new BarData(set);
-        data.setValueTextSize(10f);
-        data.setValueTypeface(mTfLight);
-        data.setDrawValues(false);
-        data.setBarWidth(0.8f);
-
-        mChart.setData(data);
-    }
 }
