@@ -1,8 +1,13 @@
 
 package com.xxmassdeveloper.mpchartexample;
 
+import android.Manifest;
+import android.content.Intent;
+import android.content.pm.PackageManager;
 import android.graphics.Color;
+import android.net.Uri;
 import android.os.Bundle;
+import androidx.core.content.ContextCompat;
 import android.util.Log;
 import android.view.Menu;
 import android.view.MenuItem;
@@ -13,7 +18,6 @@ import android.widget.TextView;
 
 import com.github.mikephil.charting.charts.BubbleChart;
 import com.github.mikephil.charting.components.Legend;
-import com.github.mikephil.charting.components.Legend.LegendPosition;
 import com.github.mikephil.charting.components.XAxis;
 import com.github.mikephil.charting.components.YAxis;
 import com.github.mikephil.charting.data.BubbleData;
@@ -33,10 +37,10 @@ import java.util.ArrayList;
 public class BubbleChartActivity extends DemoBase implements OnSeekBarChangeListener,
         OnChartValueSelectedListener {
 
-    private BubbleChart mChart;
-    private SeekBar mSeekBarX, mSeekBarY;
+    private BubbleChart chart;
+    private SeekBar seekBarX, seekBarY;
     private TextView tvX, tvY;
-    
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -44,52 +48,106 @@ public class BubbleChartActivity extends DemoBase implements OnSeekBarChangeList
                 WindowManager.LayoutParams.FLAG_FULLSCREEN);
         setContentView(R.layout.activity_bubblechart);
 
+        setTitle("BubbleChartActivity");
+
         tvX = findViewById(R.id.tvXMax);
         tvY = findViewById(R.id.tvYMax);
 
-        mSeekBarX = findViewById(R.id.seekBar1);
-        mSeekBarX.setOnSeekBarChangeListener(this);
+        seekBarX = findViewById(R.id.seekBar1);
+        seekBarX.setOnSeekBarChangeListener(this);
 
-        mSeekBarY = findViewById(R.id.seekBar2);
-        mSeekBarY.setOnSeekBarChangeListener(this);
+        seekBarY = findViewById(R.id.seekBar2);
+        seekBarY.setOnSeekBarChangeListener(this);
 
-        mChart = findViewById(R.id.chart1);
-        mChart.getDescription().setEnabled(false);
+        chart = findViewById(R.id.chart1);
+        chart.getDescription().setEnabled(false);
 
-        mChart.setOnChartValueSelectedListener(this);
+        chart.setOnChartValueSelectedListener(this);
 
-        mChart.setDrawGridBackground(false);
+        chart.setDrawGridBackground(false);
 
-        mChart.setTouchEnabled(true);
+        chart.setTouchEnabled(true);
 
         // enable scaling and dragging
-        mChart.setDragEnabled(true);
-        mChart.setScaleEnabled(true);
+        chart.setDragEnabled(true);
+        chart.setScaleEnabled(true);
 
-        mChart.setMaxVisibleValueCount(200);
-        mChart.setPinchZoom(true);
+        chart.setMaxVisibleValueCount(200);
+        chart.setPinchZoom(true);
 
-        mSeekBarX.setProgress(10);
-        mSeekBarY.setProgress(50);
+        seekBarX.setProgress(10);
+        seekBarY.setProgress(50);
 
-        Legend l = mChart.getLegend();
+        Legend l = chart.getLegend();
         l.setVerticalAlignment(Legend.LegendVerticalAlignment.TOP);
         l.setHorizontalAlignment(Legend.LegendHorizontalAlignment.RIGHT);
         l.setOrientation(Legend.LegendOrientation.VERTICAL);
         l.setDrawInside(false);
-        l.setTypeface(mTfLight);
+        l.setTypeface(tfLight);
 
-        YAxis yl = mChart.getAxisLeft();
-        yl.setTypeface(mTfLight);
+        YAxis yl = chart.getAxisLeft();
+        yl.setTypeface(tfLight);
         yl.setSpaceTop(30f);
         yl.setSpaceBottom(30f);
         yl.setDrawZeroLine(false);
-        
-        mChart.getAxisRight().setEnabled(false);
 
-        XAxis xl = mChart.getXAxis();
+        chart.getAxisRight().setEnabled(false);
+
+        XAxis xl = chart.getXAxis();
         xl.setPosition(XAxis.XAxisPosition.BOTTOM);
-        xl.setTypeface(mTfLight);
+        xl.setTypeface(tfLight);
+    }
+
+    @Override
+    public void onProgressChanged(SeekBar seekBar, int progress, boolean fromUser) {
+
+        int count = seekBarX.getProgress();
+        int range = seekBarY.getProgress();
+
+        tvX.setText(String.valueOf(count));
+        tvY.setText(String.valueOf(range));
+
+        ArrayList<BubbleEntry> values1 = new ArrayList<>();
+        ArrayList<BubbleEntry> values2 = new ArrayList<>();
+        ArrayList<BubbleEntry> values3 = new ArrayList<>();
+
+        for (int i = 0; i < count; i++) {
+            values1.add(new BubbleEntry(i, (float) (Math.random() * range), (float) (Math.random() * range), getResources().getDrawable(R.drawable.star)));
+            values2.add(new BubbleEntry(i, (float) (Math.random() * range), (float) (Math.random() * range), getResources().getDrawable(R.drawable.star)));
+            values3.add(new BubbleEntry(i, (float) (Math.random() * range), (float) (Math.random() * range)));
+        }
+
+        // create a dataset and give it a type
+        BubbleDataSet set1 = new BubbleDataSet(values1, "DS 1");
+        set1.setDrawIcons(false);
+        set1.setColor(ColorTemplate.COLORFUL_COLORS[0], 130);
+        set1.setDrawValues(true);
+
+        BubbleDataSet set2 = new BubbleDataSet(values2, "DS 2");
+        set2.setDrawIcons(false);
+        set2.setIconsOffset(new MPPointF(0, 15));
+        set2.setColor(ColorTemplate.COLORFUL_COLORS[1], 130);
+        set2.setDrawValues(true);
+
+        BubbleDataSet set3 = new BubbleDataSet(values3, "DS 3");
+        set3.setColor(ColorTemplate.COLORFUL_COLORS[2], 130);
+        set3.setDrawValues(true);
+
+        ArrayList<IBubbleDataSet> dataSets = new ArrayList<>();
+        dataSets.add(set1); // add the data sets
+        dataSets.add(set2);
+        dataSets.add(set3);
+
+        // create a data object with the data sets
+        BubbleData data = new BubbleData(dataSets);
+        data.setDrawValues(false);
+        data.setValueTypeface(tfLight);
+        data.setValueTextSize(8f);
+        data.setValueTextColor(Color.WHITE);
+        data.setHighlightCircleWidth(1.5f);
+
+        chart.setData(data);
+        chart.invalidate();
     }
 
     @Override
@@ -102,56 +160,65 @@ public class BubbleChartActivity extends DemoBase implements OnSeekBarChangeList
     public boolean onOptionsItemSelected(MenuItem item) {
 
         switch (item.getItemId()) {
+            case R.id.viewGithub: {
+                Intent i = new Intent(Intent.ACTION_VIEW);
+                i.setData(Uri.parse("https://github.com/PhilJay/MPAndroidChart/blob/master/MPChartExample/src/com/xxmassdeveloper/mpchartexample/BubbleChartActivity.java"));
+                startActivity(i);
+                break;
+            }
             case R.id.actionToggleValues: {
-                for (IDataSet set : mChart.getData().getDataSets())
+                for (IDataSet set : chart.getData().getDataSets())
                     set.setDrawValues(!set.isDrawValuesEnabled());
 
-                mChart.invalidate();
+                chart.invalidate();
                 break;
             }
             case R.id.actionToggleIcons: {
-                for (IDataSet set : mChart.getData().getDataSets())
+                for (IDataSet set : chart.getData().getDataSets())
                     set.setDrawIcons(!set.isDrawIconsEnabled());
 
-                mChart.invalidate();
+                chart.invalidate();
                 break;
             }
             case R.id.actionToggleHighlight: {
-                if(mChart.getData() != null) {
-                    mChart.getData().setHighlightEnabled(!mChart.getData().isHighlightEnabled());
-                    mChart.invalidate();
+                if(chart.getData() != null) {
+                    chart.getData().setHighlightEnabled(!chart.getData().isHighlightEnabled());
+                    chart.invalidate();
                 }
                 break;
             }
             case R.id.actionTogglePinch: {
-                if (mChart.isPinchZoomEnabled())
-                    mChart.setPinchZoom(false);
+                if (chart.isPinchZoomEnabled())
+                    chart.setPinchZoom(false);
                 else
-                    mChart.setPinchZoom(true);
+                    chart.setPinchZoom(true);
 
-                mChart.invalidate();
+                chart.invalidate();
                 break;
             }
             case R.id.actionToggleAutoScaleMinMax: {
-                mChart.setAutoScaleMinMaxEnabled(!mChart.isAutoScaleMinMaxEnabled());
-                mChart.notifyDataSetChanged();
+                chart.setAutoScaleMinMaxEnabled(!chart.isAutoScaleMinMaxEnabled());
+                chart.notifyDataSetChanged();
                 break;
             }
             case R.id.actionSave: {
-                mChart.saveToPath("title" + System.currentTimeMillis(), "");
+                if (ContextCompat.checkSelfPermission(this, Manifest.permission.WRITE_EXTERNAL_STORAGE) == PackageManager.PERMISSION_GRANTED) {
+                    saveToGallery();
+                } else {
+                    requestStoragePermission(chart);
+                }
                 break;
             }
             case R.id.animateX: {
-                mChart.animateX(3000);
+                chart.animateX(2000);
                 break;
             }
             case R.id.animateY: {
-                mChart.animateY(3000);
+                chart.animateY(2000);
                 break;
             }
             case R.id.animateXY: {
-
-                mChart.animateXY(3000, 3000);
+                chart.animateXY(2000, 2000);
                 break;
             }
         }
@@ -159,70 +226,8 @@ public class BubbleChartActivity extends DemoBase implements OnSeekBarChangeList
     }
 
     @Override
-    public void onProgressChanged(SeekBar seekBar, int progress, boolean fromUser) {
-
-        int count = mSeekBarX.getProgress();
-        int range = mSeekBarY.getProgress();
-
-        tvX.setText("" + count);
-        tvY.setText("" + range);
-
-        ArrayList<BubbleEntry> yVals1 = new ArrayList<BubbleEntry>();
-        ArrayList<BubbleEntry> yVals2 = new ArrayList<BubbleEntry>();
-        ArrayList<BubbleEntry> yVals3 = new ArrayList<BubbleEntry>();
-
-        for (int i = 0; i < count; i++) {
-            float val = (float) (Math.random() * range);
-            float size = (float) (Math.random() * range);
-
-            yVals1.add(new BubbleEntry(i, val, size, getResources().getDrawable(R.drawable.star)));
-        }
-
-        for (int i = 0; i < count; i++) {
-            float val = (float) (Math.random() * range);
-            float size = (float) (Math.random() * range);
-
-            yVals2.add(new BubbleEntry(i, val, size, getResources().getDrawable(R.drawable.star)));
-        }
-
-        for (int i = 0; i < count; i++) {
-            float val = (float) (Math.random() * range);
-            float size = (float) (Math.random() * range);
-
-            yVals3.add(new BubbleEntry(i, val, size));
-        }
-
-        // create a dataset and give it a type
-        BubbleDataSet set1 = new BubbleDataSet(yVals1, "DS 1");
-        set1.setDrawIcons(false);
-        set1.setColor(ColorTemplate.COLORFUL_COLORS[0], 130);
-        set1.setDrawValues(true);
-
-        BubbleDataSet set2 = new BubbleDataSet(yVals2, "DS 2");
-        set2.setDrawIcons(false);
-        set2.setIconsOffset(new MPPointF(0, 15));
-        set2.setColor(ColorTemplate.COLORFUL_COLORS[1], 130);
-        set2.setDrawValues(true);
-
-        BubbleDataSet set3 = new BubbleDataSet(yVals3, "DS 3");
-        set3.setColor(ColorTemplate.COLORFUL_COLORS[2], 130);
-        set3.setDrawValues(true);
-
-        ArrayList<IBubbleDataSet> dataSets = new ArrayList<IBubbleDataSet>();
-        dataSets.add(set1); // add the datasets
-        dataSets.add(set2);
-        dataSets.add(set3);
-
-        // create a data object with the datasets
-        BubbleData data = new BubbleData(dataSets);
-        data.setDrawValues(false);
-        data.setValueTypeface(mTfLight);
-        data.setValueTextSize(8f);
-        data.setValueTextColor(Color.WHITE);
-        data.setHighlightCircleWidth(1.5f);
-
-        mChart.setData(data);
-        mChart.invalidate();
+    public void saveToGallery() {
+        saveToGallery(chart, "BubbleChartActivity");
     }
 
     @Override
@@ -233,20 +238,11 @@ public class BubbleChartActivity extends DemoBase implements OnSeekBarChangeList
     }
 
     @Override
-    public void onNothingSelected() {
-        // TODO Auto-generated method stub
-
-    }
+    public void onNothingSelected() {}
 
     @Override
-    public void onStartTrackingTouch(SeekBar seekBar) {
-        // TODO Auto-generated method stub
-
-    }
+    public void onStartTrackingTouch(SeekBar seekBar) {}
 
     @Override
-    public void onStopTrackingTouch(SeekBar seekBar) {
-        // TODO Auto-generated method stub
-
-    }
+    public void onStopTrackingTouch(SeekBar seekBar) {}
 }
