@@ -1,16 +1,20 @@
 
 package com.xxmassdeveloper.mpchartexample;
 
+import android.Manifest;
+import android.content.Intent;
+import android.content.pm.PackageManager;
 import android.graphics.Color;
 import android.graphics.Paint;
+import android.net.Uri;
 import android.os.Bundle;
+import androidx.core.content.ContextCompat;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.WindowManager;
 import android.widget.SeekBar;
 import android.widget.SeekBar.OnSeekBarChangeListener;
 import android.widget.TextView;
-import android.widget.Toast;
 
 import com.github.mikephil.charting.charts.CandleStickChart;
 import com.github.mikephil.charting.components.XAxis;
@@ -28,8 +32,8 @@ import java.util.ArrayList;
 
 public class CandleStickChartActivity extends DemoBase implements OnSeekBarChangeListener {
 
-    private CandleStickChart mChart;
-    private SeekBar mSeekBarX, mSeekBarY;
+    private CandleStickChart chart;
+    private SeekBar seekBarX, seekBarY;
     private TextView tvX, tvY;
 
     @Override
@@ -39,154 +43,77 @@ public class CandleStickChartActivity extends DemoBase implements OnSeekBarChang
                 WindowManager.LayoutParams.FLAG_FULLSCREEN);
         setContentView(R.layout.activity_candlechart);
 
+        setTitle("CandleStickChartActivity");
+
         tvX = findViewById(R.id.tvXMax);
         tvY = findViewById(R.id.tvYMax);
 
-        mSeekBarX = findViewById(R.id.seekBar1);
-        mSeekBarX.setOnSeekBarChangeListener(this);
+        seekBarX = findViewById(R.id.seekBar1);
+        seekBarX.setOnSeekBarChangeListener(this);
 
-        mSeekBarY = findViewById(R.id.seekBar2);
-        mSeekBarY.setOnSeekBarChangeListener(this);
+        seekBarY = findViewById(R.id.seekBar2);
+        seekBarY.setOnSeekBarChangeListener(this);
 
-        mChart = findViewById(R.id.chart1);
-        mChart.setBackgroundColor(Color.WHITE);
+        chart = findViewById(R.id.chart1);
+        chart.setBackgroundColor(Color.WHITE);
 
-        mChart.getDescription().setEnabled(false);
+        chart.getDescription().setEnabled(false);
 
         // if more than 60 entries are displayed in the chart, no values will be
         // drawn
-        mChart.setMaxVisibleValueCount(60);
+        chart.setMaxVisibleValueCount(60);
 
         // scaling can now only be done on x- and y-axis separately
-        mChart.setPinchZoom(false);
+        chart.setPinchZoom(false);
 
-        mChart.setDrawGridBackground(false);
+        chart.setDrawGridBackground(false);
 
-        XAxis xAxis = mChart.getXAxis();
+        XAxis xAxis = chart.getXAxis();
         xAxis.setPosition(XAxisPosition.BOTTOM);
         xAxis.setDrawGridLines(false);
 
-        YAxis leftAxis = mChart.getAxisLeft();  
+        YAxis leftAxis = chart.getAxisLeft();
 //        leftAxis.setEnabled(false);
         leftAxis.setLabelCount(7, false);
         leftAxis.setDrawGridLines(false);
         leftAxis.setDrawAxisLine(false);
-        
-        YAxis rightAxis = mChart.getAxisRight();
+
+        YAxis rightAxis = chart.getAxisRight();
         rightAxis.setEnabled(false);
 //        rightAxis.setStartAtZero(false);
 
         // setting data
-        mSeekBarX.setProgress(40);
-        mSeekBarY.setProgress(100);
-        
-        mChart.getLegend().setEnabled(false);
-    }
+        seekBarX.setProgress(40);
+        seekBarY.setProgress(100);
 
-    @Override
-    public boolean onCreateOptionsMenu(Menu menu) {
-        getMenuInflater().inflate(R.menu.candle, menu);
-        return true;
-    }
-
-    @Override
-    public boolean onOptionsItemSelected(MenuItem item) {
-
-        switch (item.getItemId()) {
-            case R.id.actionToggleValues: {
-                for (IDataSet set : mChart.getData().getDataSets())
-                    set.setDrawValues(!set.isDrawValuesEnabled());
-
-                mChart.invalidate();
-                break;
-            }
-            case R.id.actionToggleIcons: {
-                for (IDataSet set : mChart.getData().getDataSets())
-                    set.setDrawIcons(!set.isDrawIconsEnabled());
-
-                mChart.invalidate();
-                break;
-            }
-            case R.id.actionToggleHighlight: {
-                if(mChart.getData() != null) {
-                    mChart.getData().setHighlightEnabled(!mChart.getData().isHighlightEnabled());
-                    mChart.invalidate();
-                }
-                break;
-            }
-            case R.id.actionTogglePinch: {
-                if (mChart.isPinchZoomEnabled())
-                    mChart.setPinchZoom(false);
-                else
-                    mChart.setPinchZoom(true);
-
-                mChart.invalidate();
-                break;
-            }
-            case R.id.actionToggleAutoScaleMinMax: {
-                mChart.setAutoScaleMinMaxEnabled(!mChart.isAutoScaleMinMaxEnabled());
-                mChart.notifyDataSetChanged();
-                break;
-            }
-            case R.id.actionToggleMakeShadowSameColorAsCandle: {
-                for (ICandleDataSet set : mChart.getData().getDataSets()) {
-                   //TODO: set.setShadowColorSameAsCandle(!set.getShadowColorSameAsCandle());
-                }
-
-                mChart.invalidate();
-                break;
-            }
-            case R.id.animateX: {
-                mChart.animateX(3000);
-                break;
-            }
-            case R.id.animateY: {
-                mChart.animateY(3000);
-                break;
-            }
-            case R.id.animateXY: {
-
-                mChart.animateXY(3000, 3000);
-                break;
-            }
-            case R.id.actionSave: {
-                if (mChart.saveToGallery("title" + System.currentTimeMillis(), 50)) {
-                    Toast.makeText(getApplicationContext(), "Saving SUCCESSFUL!",
-                            Toast.LENGTH_SHORT).show();
-                } else
-                    Toast.makeText(getApplicationContext(), "Saving FAILED!", Toast.LENGTH_SHORT)
-                            .show();
-                break;
-            }
-        }
-        return true;
+        chart.getLegend().setEnabled(false);
     }
 
     @Override
     public void onProgressChanged(SeekBar seekBar, int progress, boolean fromUser) {
-        
-        int prog = (mSeekBarX.getProgress() + 1);
 
-        tvX.setText("" + prog);
-        tvY.setText("" + (mSeekBarY.getProgress()));
-        
-        mChart.resetTracking();
+        progress = (seekBarX.getProgress());
 
-        ArrayList<CandleEntry> yVals1 = new ArrayList<CandleEntry>();
+        tvX.setText(String.valueOf(progress));
+        tvY.setText(String.valueOf(seekBarY.getProgress()));
 
-        for (int i = 0; i < prog; i++) {
-            float mult = (mSeekBarY.getProgress() + 1);
-            float val = (float) (Math.random() * 40) + mult;
-            
+        chart.resetTracking();
+
+        ArrayList<CandleEntry> values = new ArrayList<>();
+
+        for (int i = 0; i < progress; i++) {
+            float multi = (seekBarY.getProgress() + 1);
+            float val = (float) (Math.random() * 40) + multi;
+
             float high = (float) (Math.random() * 9) + 8f;
             float low = (float) (Math.random() * 9) + 8f;
-            
+
             float open = (float) (Math.random() * 6) + 1f;
             float close = (float) (Math.random() * 6) + 1f;
 
             boolean even = i % 2 == 0;
 
-            yVals1.add(new CandleEntry(
+            values.add(new CandleEntry(
                     i, val + high,
                     val - low,
                     even ? val + open : val - open,
@@ -195,7 +122,7 @@ public class CandleStickChartActivity extends DemoBase implements OnSeekBarChang
             ));
         }
 
-        CandleDataSet set1 = new CandleDataSet(yVals1, "Data Set");
+        CandleDataSet set1 = new CandleDataSet(values, "Data Set");
 
         set1.setDrawIcons(false);
         set1.setAxisDependency(AxisDependency.LEFT);
@@ -210,20 +137,102 @@ public class CandleStickChartActivity extends DemoBase implements OnSeekBarChang
         //set1.setHighlightLineWidth(1f);
 
         CandleData data = new CandleData(set1);
-        
-        mChart.setData(data);
-        mChart.invalidate();
+
+        chart.setData(data);
+        chart.invalidate();
     }
 
     @Override
-    public void onStartTrackingTouch(SeekBar seekBar) {
-        // TODO Auto-generated method stub
-
+    public boolean onCreateOptionsMenu(Menu menu) {
+        getMenuInflater().inflate(R.menu.candle, menu);
+        return true;
     }
 
     @Override
-    public void onStopTrackingTouch(SeekBar seekBar) {
-        // TODO Auto-generated method stub
+    public boolean onOptionsItemSelected(MenuItem item) {
 
+        switch (item.getItemId()) {
+            case R.id.viewGithub: {
+                Intent i = new Intent(Intent.ACTION_VIEW);
+                i.setData(Uri.parse("https://github.com/PhilJay/MPAndroidChart/blob/master/MPChartExample/src/com/xxmassdeveloper/mpchartexample/CandleStickChartActivity.java"));
+                startActivity(i);
+                break;
+            }
+            case R.id.actionToggleValues: {
+                for (IDataSet set : chart.getData().getDataSets())
+                    set.setDrawValues(!set.isDrawValuesEnabled());
+
+                chart.invalidate();
+                break;
+            }
+            case R.id.actionToggleIcons: {
+                for (IDataSet set : chart.getData().getDataSets())
+                    set.setDrawIcons(!set.isDrawIconsEnabled());
+
+                chart.invalidate();
+                break;
+            }
+            case R.id.actionToggleHighlight: {
+                if(chart.getData() != null) {
+                    chart.getData().setHighlightEnabled(!chart.getData().isHighlightEnabled());
+                    chart.invalidate();
+                }
+                break;
+            }
+            case R.id.actionTogglePinch: {
+                if (chart.isPinchZoomEnabled())
+                    chart.setPinchZoom(false);
+                else
+                    chart.setPinchZoom(true);
+
+                chart.invalidate();
+                break;
+            }
+            case R.id.actionToggleAutoScaleMinMax: {
+                chart.setAutoScaleMinMaxEnabled(!chart.isAutoScaleMinMaxEnabled());
+                chart.notifyDataSetChanged();
+                break;
+            }
+            case R.id.actionToggleMakeShadowSameColorAsCandle: {
+                for (ICandleDataSet set : chart.getData().getDataSets()) {
+                    ((CandleDataSet) set).setShadowColorSameAsCandle(!set.getShadowColorSameAsCandle());
+                }
+
+                chart.invalidate();
+                break;
+            }
+            case R.id.animateX: {
+                chart.animateX(2000);
+                break;
+            }
+            case R.id.animateY: {
+                chart.animateY(2000);
+                break;
+            }
+            case R.id.animateXY: {
+                chart.animateXY(2000, 2000);
+                break;
+            }
+            case R.id.actionSave: {
+                if (ContextCompat.checkSelfPermission(this, Manifest.permission.WRITE_EXTERNAL_STORAGE) == PackageManager.PERMISSION_GRANTED) {
+                    saveToGallery();
+                } else {
+                    requestStoragePermission(chart);
+                }
+                break;
+            }
+        }
+        return true;
     }
+
+    @Override
+    public void saveToGallery() {
+        saveToGallery(chart, "CandleStickChartActivity");
+    }
+
+    @Override
+    public void onStartTrackingTouch(SeekBar seekBar) {}
+
+    @Override
+    public void onStopTrackingTouch(SeekBar seekBar) {}
 }
