@@ -1,8 +1,13 @@
 
 package com.xxmassdeveloper.mpchartexample;
 
+import android.Manifest;
+import android.content.Intent;
+import android.content.pm.PackageManager;
 import android.graphics.Color;
+import android.net.Uri;
 import android.os.Bundle;
+import androidx.core.content.ContextCompat;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.WindowManager;
@@ -23,7 +28,7 @@ import java.util.ArrayList;
 
 public class DynamicalAddingActivity extends DemoBase implements OnChartValueSelectedListener {
 
-    private LineChart mChart;
+    private LineChart chart;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -32,24 +37,30 @@ public class DynamicalAddingActivity extends DemoBase implements OnChartValueSel
                 WindowManager.LayoutParams.FLAG_FULLSCREEN);
         setContentView(R.layout.activity_linechart_noseekbar);
 
-        mChart = findViewById(R.id.chart1);
-        mChart.setOnChartValueSelectedListener(this);
-        mChart.setDrawGridBackground(false);
-        mChart.getDescription().setEnabled(false);
+        setTitle("DynamicalAddingActivity");
 
-        // add an empty data object
-        mChart.setData(new LineData());
-//        mChart.getXAxis().setDrawLabels(false);
-//        mChart.getXAxis().setDrawGridLines(false);
+        chart = findViewById(R.id.chart1);
+        chart.setOnChartValueSelectedListener(this);
+        chart.setDrawGridBackground(false);
+        chart.getDescription().setEnabled(false);
+        chart.setNoDataText("No chart data available. Use the menu to add entries and data sets!");
 
-        mChart.invalidate();
+//        chart.getXAxis().setDrawLabels(false);
+//        chart.getXAxis().setDrawGridLines(false);
+
+        chart.invalidate();
     }
 
-    int[] mColors = ColorTemplate.VORDIPLOM_COLORS;
+    private final int[] colors = ColorTemplate.VORDIPLOM_COLORS;
 
     private void addEntry() {
 
-        LineData data = mChart.getData();
+        LineData data = chart.getData();
+
+        if (data == null) {
+            data = new LineData();
+            chart.setData(data);
+        }
 
         ILineDataSet set = data.getDataSetByIndex(0);
         // set.addEntry(...); // can be called as well
@@ -61,25 +72,26 @@ public class DynamicalAddingActivity extends DemoBase implements OnChartValueSel
 
         // choose a random dataSet
         int randomDataSetIndex = (int) (Math.random() * data.getDataSetCount());
-        float yValue = (float) (Math.random() * 10) + 50f;
+        ILineDataSet randomSet = data.getDataSetByIndex(randomDataSetIndex);
+        float value = (float) (Math.random() * 50) + 50f * (randomDataSetIndex + 1);
 
-        data.addEntry(new Entry(data.getDataSetByIndex(randomDataSetIndex).getEntryCount(), yValue), randomDataSetIndex);
+        data.addEntry(new Entry(randomSet.getEntryCount(), value), randomDataSetIndex);
         data.notifyDataChanged();
 
         // let the chart know it's data has changed
-        mChart.notifyDataSetChanged();
+        chart.notifyDataSetChanged();
 
-        mChart.setVisibleXRangeMaximum(6);
-        //mChart.setVisibleYRangeMaximum(15, AxisDependency.LEFT);
-//            
+        chart.setVisibleXRangeMaximum(6);
+        //chart.setVisibleYRangeMaximum(15, AxisDependency.LEFT);
+//
 //            // this automatically refreshes the chart (calls invalidate())
-        mChart.moveViewTo(data.getEntryCount() - 7, 50f, AxisDependency.LEFT);
+        chart.moveViewTo(data.getEntryCount() - 7, 50f, AxisDependency.LEFT);
 
     }
 
     private void removeLastEntry() {
 
-        LineData data = mChart.getData();
+        LineData data = chart.getData();
 
         if (data != null) {
 
@@ -93,31 +105,33 @@ public class DynamicalAddingActivity extends DemoBase implements OnChartValueSel
                 // or remove by index
                 // mData.removeEntryByXValue(xIndex, dataSetIndex);
                 data.notifyDataChanged();
-                mChart.notifyDataSetChanged();
-                mChart.invalidate();
+                chart.notifyDataSetChanged();
+                chart.invalidate();
             }
         }
     }
 
     private void addDataSet() {
 
-        LineData data = mChart.getData();
+        LineData data = chart.getData();
 
-        if (data != null) {
-
+        if (data == null) {
+            chart.setData(new LineData());
+        } else {
             int count = (data.getDataSetCount() + 1);
+            int amount = data.getDataSetByIndex(0).getEntryCount();
 
-            ArrayList<Entry> yVals = new ArrayList<Entry>();
+            ArrayList<Entry> values = new ArrayList<>();
 
-            for (int i = 0; i < data.getEntryCount(); i++) {
-                yVals.add(new Entry(i, (float) (Math.random() * 50f) + 50f * count));
+            for (int i = 0; i < amount; i++) {
+                values.add(new Entry(i, (float) (Math.random() * 50f) + 50f * count));
             }
 
-            LineDataSet set = new LineDataSet(yVals, "DataSet " + count);
+            LineDataSet set = new LineDataSet(values, "DataSet " + count);
             set.setLineWidth(2.5f);
             set.setCircleRadius(4.5f);
 
-            int color = mColors[count % mColors.length];
+            int color = colors[count % colors.length];
 
             set.setColor(color);
             set.setCircleColor(color);
@@ -127,72 +141,22 @@ public class DynamicalAddingActivity extends DemoBase implements OnChartValueSel
 
             data.addDataSet(set);
             data.notifyDataChanged();
-            mChart.notifyDataSetChanged();
-            mChart.invalidate();
+            chart.notifyDataSetChanged();
+            chart.invalidate();
         }
     }
 
     private void removeDataSet() {
 
-        LineData data = mChart.getData();
+        LineData data = chart.getData();
 
         if (data != null) {
 
             data.removeDataSet(data.getDataSetByIndex(data.getDataSetCount() - 1));
 
-            mChart.notifyDataSetChanged();
-            mChart.invalidate();
+            chart.notifyDataSetChanged();
+            chart.invalidate();
         }
-    }
-
-    @Override
-    public void onValueSelected(Entry e, Highlight h) {
-        Toast.makeText(this, e.toString(), Toast.LENGTH_SHORT).show();
-    }
-
-    @Override
-    public void onNothingSelected() {
-
-    }
-
-    @Override
-    public boolean onCreateOptionsMenu(Menu menu) {
-        getMenuInflater().inflate(R.menu.dynamical, menu);
-        return true;
-    }
-
-    @Override
-    public boolean onOptionsItemSelected(MenuItem item) {
-
-        switch (item.getItemId()) {
-            case R.id.actionAddEntry:
-                addEntry();
-                Toast.makeText(this, "Entry added!", Toast.LENGTH_SHORT).show();
-                break;
-            case R.id.actionRemoveEntry:
-                removeLastEntry();
-                Toast.makeText(this, "Entry removed!", Toast.LENGTH_SHORT).show();
-                break;
-            case R.id.actionAddDataSet:
-                addDataSet();
-                Toast.makeText(this, "DataSet added!", Toast.LENGTH_SHORT).show();
-                break;
-            case R.id.actionRemoveDataSet:
-                removeDataSet();
-                Toast.makeText(this, "DataSet removed!", Toast.LENGTH_SHORT).show();
-                break;
-            case R.id.actionAddEmptyLineData:
-                mChart.setData(new LineData());
-                mChart.invalidate();
-                Toast.makeText(this, "Empty data added!", Toast.LENGTH_SHORT).show();
-                break;
-            case R.id.actionClear:
-                mChart.clear();
-                Toast.makeText(this, "Chart cleared!", Toast.LENGTH_SHORT).show();
-                break;
-        }
-
-        return true;
     }
 
     private LineDataSet createSet() {
@@ -207,5 +171,72 @@ public class DynamicalAddingActivity extends DemoBase implements OnChartValueSel
         set.setValueTextSize(10f);
 
         return set;
+    }
+
+    @Override
+    public void onValueSelected(Entry e, Highlight h) {
+        Toast.makeText(this, e.toString(), Toast.LENGTH_SHORT).show();
+    }
+
+    @Override
+    public void onNothingSelected() {}
+
+    @Override
+    public boolean onCreateOptionsMenu(Menu menu) {
+        getMenuInflater().inflate(R.menu.dynamical, menu);
+        return true;
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+
+        switch (item.getItemId()) {
+            case R.id.viewGithub: {
+                Intent i = new Intent(Intent.ACTION_VIEW);
+                i.setData(Uri.parse("https://github.com/PhilJay/MPAndroidChart/blob/master/MPChartExample/src/com/xxmassdeveloper/mpchartexample/DynamicalAddingActivity.java"));
+                startActivity(i);
+                break;
+            }
+            case R.id.actionAddEntry: {
+                addEntry();
+                Toast.makeText(this, "Entry added!", Toast.LENGTH_SHORT).show();
+                break;
+            }
+            case R.id.actionRemoveEntry: {
+                removeLastEntry();
+                Toast.makeText(this, "Entry removed!", Toast.LENGTH_SHORT).show();
+                break;
+            }
+            case R.id.actionAddDataSet: {
+                addDataSet();
+                Toast.makeText(this, "DataSet added!", Toast.LENGTH_SHORT).show();
+                break;
+            }
+            case R.id.actionRemoveDataSet: {
+                removeDataSet();
+                Toast.makeText(this, "DataSet removed!", Toast.LENGTH_SHORT).show();
+                break;
+            }
+            case R.id.actionClear: {
+                chart.clear();
+                Toast.makeText(this, "Chart cleared!", Toast.LENGTH_SHORT).show();
+                break;
+            }
+            case R.id.actionSave: {
+                if (ContextCompat.checkSelfPermission(this, Manifest.permission.WRITE_EXTERNAL_STORAGE) == PackageManager.PERMISSION_GRANTED) {
+                    saveToGallery();
+                } else {
+                    requestStoragePermission(chart);
+                }
+                break;
+            }
+        }
+
+        return true;
+    }
+
+    @Override
+    public void saveToGallery() {
+        saveToGallery(chart, "DynamicalAddingActivity");
     }
 }
