@@ -1,9 +1,13 @@
 
 package com.xxmassdeveloper.mpchartexample;
 
-import android.annotation.SuppressLint;
+import android.Manifest;
+import android.content.Intent;
+import android.content.pm.PackageManager;
 import android.graphics.RectF;
+import android.net.Uri;
 import android.os.Bundle;
+import androidx.core.content.ContextCompat;
 import android.util.Log;
 import android.view.Menu;
 import android.view.MenuItem;
@@ -11,11 +15,9 @@ import android.view.WindowManager;
 import android.widget.SeekBar;
 import android.widget.SeekBar.OnSeekBarChangeListener;
 import android.widget.TextView;
-import android.widget.Toast;
 
 import com.github.mikephil.charting.charts.HorizontalBarChart;
 import com.github.mikephil.charting.components.Legend;
-import com.github.mikephil.charting.components.Legend.LegendPosition;
 import com.github.mikephil.charting.components.XAxis;
 import com.github.mikephil.charting.components.XAxis.XAxisPosition;
 import com.github.mikephil.charting.components.YAxis;
@@ -35,8 +37,8 @@ import java.util.List;
 public class HorizontalBarChartActivity extends DemoBase implements OnSeekBarChangeListener,
         OnChartValueSelectedListener {
 
-    protected HorizontalBarChart mChart;
-    private SeekBar mSeekBarX, mSeekBarY;
+    private HorizontalBarChart chart;
+    private SeekBar seekBarX, seekBarY;
     private TextView tvX, tvY;
 
     @Override
@@ -46,73 +48,111 @@ public class HorizontalBarChartActivity extends DemoBase implements OnSeekBarCha
                 WindowManager.LayoutParams.FLAG_FULLSCREEN);
         setContentView(R.layout.activity_horizontalbarchart);
 
+        setTitle("HorizontalBarChartActivity");
+
         tvX = findViewById(R.id.tvXMax);
         tvY = findViewById(R.id.tvYMax);
 
-        mSeekBarX = findViewById(R.id.seekBar1);
-        mSeekBarY = findViewById(R.id.seekBar2);
+        seekBarX = findViewById(R.id.seekBar1);
+        seekBarY = findViewById(R.id.seekBar2);
 
-        mChart = findViewById(R.id.chart1);
-        mChart.setOnChartValueSelectedListener(this);
-        // mChart.setHighlightEnabled(false);
+        seekBarY.setOnSeekBarChangeListener(this);
+        seekBarX.setOnSeekBarChangeListener(this);
 
-        mChart.setDrawBarShadow(false);
+        chart = findViewById(R.id.chart1);
+        chart.setOnChartValueSelectedListener(this);
+        // chart.setHighlightEnabled(false);
 
-        mChart.setDrawValueAboveBar(true);
+        chart.setDrawBarShadow(false);
 
-        mChart.getDescription().setEnabled(false);
+        chart.setDrawValueAboveBar(true);
+
+        chart.getDescription().setEnabled(false);
 
         // if more than 60 entries are displayed in the chart, no values will be
         // drawn
-        mChart.setMaxVisibleValueCount(60);
+        chart.setMaxVisibleValueCount(60);
 
         // scaling can now only be done on x- and y-axis separately
-        mChart.setPinchZoom(false);
+        chart.setPinchZoom(false);
 
         // draw shadows for each bar that show the maximum value
-        // mChart.setDrawBarShadow(true);
+        // chart.setDrawBarShadow(true);
 
-        mChart.setDrawGridBackground(false);
+        chart.setDrawGridBackground(false);
 
-        XAxis xl = mChart.getXAxis();
+        XAxis xl = chart.getXAxis();
         xl.setPosition(XAxisPosition.BOTTOM);
-        xl.setTypeface(mTfLight);
+        xl.setTypeface(tfLight);
         xl.setDrawAxisLine(true);
         xl.setDrawGridLines(false);
         xl.setGranularity(10f);
 
-        YAxis yl = mChart.getAxisLeft();
-        yl.setTypeface(mTfLight);
+        YAxis yl = chart.getAxisLeft();
+        yl.setTypeface(tfLight);
         yl.setDrawAxisLine(true);
         yl.setDrawGridLines(true);
         yl.setAxisMinimum(0f); // this replaces setStartAtZero(true)
 //        yl.setInverted(true);
 
-        YAxis yr = mChart.getAxisRight();
-        yr.setTypeface(mTfLight);
+        YAxis yr = chart.getAxisRight();
+        yr.setTypeface(tfLight);
         yr.setDrawAxisLine(true);
         yr.setDrawGridLines(false);
         yr.setAxisMinimum(0f); // this replaces setStartAtZero(true)
 //        yr.setInverted(true);
 
         setData(12, 50);
-        mChart.setFitBars(true);
-        mChart.animateY(2500);
+        chart.setFitBars(true);
+        chart.animateY(2500);
 
         // setting data
-        mSeekBarY.setProgress(50);
-        mSeekBarX.setProgress(12);
+        seekBarY.setProgress(50);
+        seekBarX.setProgress(12);
 
-        mSeekBarY.setOnSeekBarChangeListener(this);
-        mSeekBarX.setOnSeekBarChangeListener(this);
-
-        Legend l = mChart.getLegend();
+        Legend l = chart.getLegend();
         l.setVerticalAlignment(Legend.LegendVerticalAlignment.BOTTOM);
         l.setHorizontalAlignment(Legend.LegendHorizontalAlignment.LEFT);
         l.setOrientation(Legend.LegendOrientation.HORIZONTAL);
         l.setDrawInside(false);
         l.setFormSize(8f);
         l.setXEntrySpace(4f);
+    }
+
+    private void setData(int count, float range) {
+
+        float barWidth = 9f;
+        float spaceForBar = 10f;
+        ArrayList<BarEntry> values = new ArrayList<>();
+
+        for (int i = 0; i < count; i++) {
+            float val = (float) (Math.random() * range);
+            values.add(new BarEntry(i * spaceForBar, val,
+                    getResources().getDrawable(R.drawable.star)));
+        }
+
+        BarDataSet set1;
+
+        if (chart.getData() != null &&
+                chart.getData().getDataSetCount() > 0) {
+            set1 = (BarDataSet) chart.getData().getDataSetByIndex(0);
+            set1.setValues(values);
+            chart.getData().notifyDataChanged();
+            chart.notifyDataSetChanged();
+        } else {
+            set1 = new BarDataSet(values, "DataSet 1");
+
+            set1.setDrawIcons(false);
+
+            ArrayList<IBarDataSet> dataSets = new ArrayList<>();
+            dataSets.add(set1);
+
+            BarData data = new BarData(dataSets);
+            data.setValueTextSize(10f);
+            data.setValueTypeface(tfLight);
+            data.setBarWidth(barWidth);
+            chart.setData(data);
+        }
     }
 
     @Override
@@ -125,80 +165,80 @@ public class HorizontalBarChartActivity extends DemoBase implements OnSeekBarCha
     public boolean onOptionsItemSelected(MenuItem item) {
 
         switch (item.getItemId()) {
+            case R.id.viewGithub: {
+                Intent i = new Intent(Intent.ACTION_VIEW);
+                i.setData(Uri.parse("https://github.com/PhilJay/MPAndroidChart/blob/master/MPChartExample/src/com/xxmassdeveloper/mpchartexample/HorizontalBarChartActivity.java"));
+                startActivity(i);
+                break;
+            }
             case R.id.actionToggleValues: {
-                List<IBarDataSet> sets = mChart.getData()
+                List<IBarDataSet> sets = chart.getData()
                         .getDataSets();
 
                 for (IBarDataSet iSet : sets) {
-
-                    IBarDataSet set = (BarDataSet) iSet;
-                    set.setDrawValues(!set.isDrawValuesEnabled());
+                    iSet.setDrawValues(!iSet.isDrawValuesEnabled());
                 }
 
-                mChart.invalidate();
+                chart.invalidate();
                 break;
             }
             case R.id.actionToggleIcons: {
-                List<IBarDataSet> sets = mChart.getData()
+                List<IBarDataSet> sets = chart.getData()
                         .getDataSets();
 
                 for (IBarDataSet iSet : sets) {
-
-                    IBarDataSet set = (BarDataSet) iSet;
-                    set.setDrawIcons(!set.isDrawIconsEnabled());
+                    iSet.setDrawIcons(!iSet.isDrawIconsEnabled());
                 }
 
-                mChart.invalidate();
+                chart.invalidate();
                 break;
             }
             case R.id.actionToggleHighlight: {
-                if(mChart.getData() != null) {
-                    mChart.getData().setHighlightEnabled(!mChart.getData().isHighlightEnabled());
-                    mChart.invalidate();
+                if(chart.getData() != null) {
+                    chart.getData().setHighlightEnabled(!chart.getData().isHighlightEnabled());
+                    chart.invalidate();
                 }
                 break;
             }
             case R.id.actionTogglePinch: {
-                if (mChart.isPinchZoomEnabled())
-                    mChart.setPinchZoom(false);
+                if (chart.isPinchZoomEnabled())
+                    chart.setPinchZoom(false);
                 else
-                    mChart.setPinchZoom(true);
+                    chart.setPinchZoom(true);
 
-                mChart.invalidate();
+                chart.invalidate();
                 break;
             }
             case R.id.actionToggleAutoScaleMinMax: {
-                mChart.setAutoScaleMinMaxEnabled(!mChart.isAutoScaleMinMaxEnabled());
-                mChart.notifyDataSetChanged();
+                chart.setAutoScaleMinMaxEnabled(!chart.isAutoScaleMinMaxEnabled());
+                chart.notifyDataSetChanged();
                 break;
             }
             case R.id.actionToggleBarBorders: {
-                for (IBarDataSet set : mChart.getData().getDataSets())
+                for (IBarDataSet set : chart.getData().getDataSets())
                     ((BarDataSet)set).setBarBorderWidth(set.getBarBorderWidth() == 1.f ? 0.f : 1.f);
 
-                mChart.invalidate();
+                chart.invalidate();
                 break;
             }
             case R.id.animateX: {
-                mChart.animateX(3000);
+                chart.animateX(2000);
                 break;
             }
             case R.id.animateY: {
-                mChart.animateY(3000);
+                chart.animateY(2000);
                 break;
             }
             case R.id.animateXY: {
-
-                mChart.animateXY(3000, 3000);
+                chart.animateXY(2000, 2000);
                 break;
             }
             case R.id.actionSave: {
-                if (mChart.saveToGallery("title" + System.currentTimeMillis(), 50)) {
-                    Toast.makeText(getApplicationContext(), "Saving SUCCESSFUL!",
-                            Toast.LENGTH_SHORT).show();
-                } else
-                    Toast.makeText(getApplicationContext(), "Saving FAILED!", Toast.LENGTH_SHORT)
-                            .show();
+                if (ContextCompat.checkSelfPermission(this, Manifest.permission.WRITE_EXTERNAL_STORAGE) == PackageManager.PERMISSION_GRANTED) {
+                    saveToGallery();
+                } else {
+                    requestStoragePermission(chart);
+                }
                 break;
             }
         }
@@ -208,64 +248,27 @@ public class HorizontalBarChartActivity extends DemoBase implements OnSeekBarCha
     @Override
     public void onProgressChanged(SeekBar seekBar, int progress, boolean fromUser) {
 
-        tvX.setText("" + (mSeekBarX.getProgress() + 1));
-        tvY.setText("" + (mSeekBarY.getProgress()));
+        tvX.setText(String.valueOf(seekBarX.getProgress()));
+        tvY.setText(String.valueOf(seekBarY.getProgress()));
 
-        setData(mSeekBarX.getProgress() + 1, mSeekBarY.getProgress());
-        mChart.setFitBars(true);
-        mChart.invalidate();
+        setData(seekBarX.getProgress(), seekBarY.getProgress());
+        chart.setFitBars(true);
+        chart.invalidate();
     }
 
     @Override
-    public void onStartTrackingTouch(SeekBar seekBar) {
-        // TODO Auto-generated method stub
-
+    public void saveToGallery() {
+        saveToGallery(chart, "HorizontalBarChartActivity");
     }
 
     @Override
-    public void onStopTrackingTouch(SeekBar seekBar) {
-        // TODO Auto-generated method stub
+    public void onStartTrackingTouch(SeekBar seekBar) {}
 
-    }
+    @Override
+    public void onStopTrackingTouch(SeekBar seekBar) {}
 
-    private void setData(int count, float range) {
+    private RectF mOnValueSelectedRectF = new RectF();
 
-        float barWidth = 9f;
-        float spaceForBar = 10f;
-        ArrayList<BarEntry> yVals1 = new ArrayList<BarEntry>();
-
-        for (int i = 0; i < count; i++) {
-            float val = (float) (Math.random() * range);
-            yVals1.add(new BarEntry(i * spaceForBar, val,
-                    getResources().getDrawable(R.drawable.star)));
-        }
-
-        BarDataSet set1;
-
-        if (mChart.getData() != null &&
-                mChart.getData().getDataSetCount() > 0) {
-            set1 = (BarDataSet)mChart.getData().getDataSetByIndex(0);
-            set1.setValues(yVals1);
-            mChart.getData().notifyDataChanged();
-            mChart.notifyDataSetChanged();
-        } else {
-            set1 = new BarDataSet(yVals1, "DataSet 1");
-
-            set1.setDrawIcons(false);
-
-            ArrayList<IBarDataSet> dataSets = new ArrayList<IBarDataSet>();
-            dataSets.add(set1);
-
-            BarData data = new BarData(dataSets);
-            data.setValueTextSize(10f);
-            data.setValueTypeface(mTfLight);
-            data.setBarWidth(barWidth);
-            mChart.setData(data);
-        }
-    }
-
-    protected RectF mOnValueSelectedRectF = new RectF();
-    @SuppressLint("NewApi")
     @Override
     public void onValueSelected(Entry e, Highlight h) {
 
@@ -273,9 +276,9 @@ public class HorizontalBarChartActivity extends DemoBase implements OnSeekBarCha
             return;
 
         RectF bounds = mOnValueSelectedRectF;
-        mChart.getBarBounds((BarEntry) e, bounds);
+        chart.getBarBounds((BarEntry) e, bounds);
 
-        MPPointF position = mChart.getPosition(e, mChart.getData().getDataSetByIndex(h.getDataSetIndex())
+        MPPointF position = chart.getPosition(e, chart.getData().getDataSetByIndex(h.getDataSetIndex())
                 .getAxisDependency());
 
         Log.i("bounds", bounds.toString());
@@ -285,6 +288,5 @@ public class HorizontalBarChartActivity extends DemoBase implements OnSeekBarCha
     }
 
     @Override
-    public void onNothingSelected() {
-    };
+    public void onNothingSelected() {}
 }
