@@ -1,11 +1,16 @@
 
 package com.xxmassdeveloper.mpchartexample;
 
+import android.annotation.SuppressLint;
 import android.content.Context;
+import android.content.Intent;
 import android.graphics.Color;
-import android.graphics.Typeface;
+import android.net.Uri;
 import android.os.Bundle;
+import androidx.annotation.NonNull;
 import android.view.LayoutInflater;
+import android.view.Menu;
+import android.view.MenuItem;
 import android.view.View;
 import android.view.ViewGroup;
 import android.view.WindowManager;
@@ -28,8 +33,8 @@ import java.util.List;
 
 /**
  * Demonstrates the use of charts inside a ListView. IMPORTANT: provide a
- * specific height attribute for the chart inside your listview-item
- * 
+ * specific height attribute for the chart inside your ListView item
+ *
  * @author Philipp Jahoda
  */
 public class ListViewBarChartActivity extends DemoBase {
@@ -40,10 +45,12 @@ public class ListViewBarChartActivity extends DemoBase {
         getWindow().setFlags(WindowManager.LayoutParams.FLAG_FULLSCREEN,
                 WindowManager.LayoutParams.FLAG_FULLSCREEN);
         setContentView(R.layout.activity_listview_chart);
-        
+
+        setTitle("ListViewBarChartActivity");
+
         ListView lv = findViewById(R.id.listView1);
 
-        ArrayList<BarData> list = new ArrayList<BarData>();
+        ArrayList<BarData> list = new ArrayList<>();
 
         // 20 items
         for (int i = 0; i < 20; i++) {
@@ -56,16 +63,18 @@ public class ListViewBarChartActivity extends DemoBase {
 
     private class ChartDataAdapter extends ArrayAdapter<BarData> {
 
-        public ChartDataAdapter(Context context, List<BarData> objects) {
+        ChartDataAdapter(Context context, List<BarData> objects) {
             super(context, 0, objects);
         }
 
+        @SuppressLint("InflateParams")
+        @NonNull
         @Override
-        public View getView(int position, View convertView, ViewGroup parent) {
+        public View getView(int position, View convertView, @NonNull ViewGroup parent) {
 
             BarData data = getItem(position);
 
-            ViewHolder holder = null;
+            ViewHolder holder;
 
             if (convertView == null) {
 
@@ -82,30 +91,32 @@ public class ListViewBarChartActivity extends DemoBase {
             }
 
             // apply styling
-            data.setValueTypeface(mTfLight);
-            data.setValueTextColor(Color.BLACK);
+            if (data != null) {
+                data.setValueTypeface(tfLight);
+                data.setValueTextColor(Color.BLACK);
+            }
             holder.chart.getDescription().setEnabled(false);
             holder.chart.setDrawGridBackground(false);
 
             XAxis xAxis = holder.chart.getXAxis();
             xAxis.setPosition(XAxisPosition.BOTTOM);
-            xAxis.setTypeface(mTfLight);
+            xAxis.setTypeface(tfLight);
             xAxis.setDrawGridLines(false);
-            
+
             YAxis leftAxis = holder.chart.getAxisLeft();
-            leftAxis.setTypeface(mTfLight);
+            leftAxis.setTypeface(tfLight);
             leftAxis.setLabelCount(5, false);
             leftAxis.setSpaceTop(15f);
-            
+
             YAxis rightAxis = holder.chart.getAxisRight();
-            rightAxis.setTypeface(mTfLight);
+            rightAxis.setTypeface(tfLight);
             rightAxis.setLabelCount(5, false);
             rightAxis.setSpaceTop(15f);
 
             // set data
             holder.chart.setData(data);
             holder.chart.setFitBars(true);
-            
+
             // do not forget to refresh the chart
 //            holder.chart.invalidate();
             holder.chart.animateY(700);
@@ -121,12 +132,12 @@ public class ListViewBarChartActivity extends DemoBase {
 
     /**
      * generates a random ChartData object with just one DataSet
-     * 
-     * @return
+     *
+     * @return Bar data
      */
     private BarData generateData(int cnt) {
 
-        ArrayList<BarEntry> entries = new ArrayList<BarEntry>();
+        ArrayList<BarEntry> entries = new ArrayList<>();
 
         for (int i = 0; i < 12; i++) {
             entries.add(new BarEntry(i, (float) (Math.random() * 70) + 30));
@@ -135,12 +146,36 @@ public class ListViewBarChartActivity extends DemoBase {
         BarDataSet d = new BarDataSet(entries, "New DataSet " + cnt);
         d.setColors(ColorTemplate.VORDIPLOM_COLORS);
         d.setBarShadowColor(Color.rgb(203, 203, 203));
-        
-        ArrayList<IBarDataSet> sets = new ArrayList<IBarDataSet>();
+
+        ArrayList<IBarDataSet> sets = new ArrayList<>();
         sets.add(d);
-        
+
         BarData cd = new BarData(sets);
         cd.setBarWidth(0.9f);
         return cd;
     }
+
+    @Override
+    public boolean onCreateOptionsMenu(Menu menu) {
+        getMenuInflater().inflate(R.menu.only_github, menu);
+        return true;
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+
+        switch (item.getItemId()) {
+            case R.id.viewGithub: {
+                Intent i = new Intent(Intent.ACTION_VIEW);
+                i.setData(Uri.parse("https://github.com/PhilJay/MPAndroidChart/blob/master/MPChartExample/src/com/xxmassdeveloper/mpchartexample/ListViewBarChartActivity.java"));
+                startActivity(i);
+                break;
+            }
+        }
+
+        return true;
+    }
+
+    @Override
+    public void saveToGallery() { /* Intentionally left empty */ }
 }
