@@ -1,13 +1,16 @@
 
 package com.xxmassdeveloper.mpchartexample;
 
+import android.Manifest;
+import android.content.Intent;
+import android.content.pm.PackageManager;
 import android.graphics.Color;
+import android.net.Uri;
 import android.os.Bundle;
+import androidx.core.content.ContextCompat;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.WindowManager;
-import android.widget.TextView;
-import android.widget.Toast;
 
 import com.github.mikephil.charting.animation.Easing;
 import com.github.mikephil.charting.charts.RadarChart;
@@ -29,49 +32,46 @@ import java.util.ArrayList;
 
 public class RadarChartActivity extends DemoBase {
 
-    private RadarChart mChart;
+    private RadarChart chart;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         getWindow().setFlags(WindowManager.LayoutParams.FLAG_FULLSCREEN,
                 WindowManager.LayoutParams.FLAG_FULLSCREEN);
-        setContentView(R.layout.activity_radarchart_noseekbar);
+        setContentView(R.layout.activity_radarchart);
 
-        TextView tv = findViewById(R.id.textView);
-        tv.setTypeface(mTfLight);
-        tv.setTextColor(Color.WHITE);
-        tv.setBackgroundColor(Color.rgb(60, 65, 82));
+        setTitle("RadarChartActivity");
 
-        mChart = findViewById(R.id.chart1);
-        mChart.setBackgroundColor(Color.rgb(60, 65, 82));
+        chart = findViewById(R.id.chart1);
+        chart.setBackgroundColor(Color.rgb(60, 65, 82));
 
-        mChart.getDescription().setEnabled(false);
+        chart.getDescription().setEnabled(false);
 
-        mChart.setWebLineWidth(1f);
-        mChart.setWebColor(Color.LTGRAY);
-        mChart.setWebLineWidthInner(1f);
-        mChart.setWebColorInner(Color.LTGRAY);
-        mChart.setWebAlpha(100);
+        chart.setWebLineWidth(1f);
+        chart.setWebColor(Color.LTGRAY);
+        chart.setWebLineWidthInner(1f);
+        chart.setWebColorInner(Color.LTGRAY);
+        chart.setWebAlpha(100);
 
         // create a custom MarkerView (extend MarkerView) and specify the layout
         // to use for it
         MarkerView mv = new RadarMarkerView(this, R.layout.radar_markerview);
-        mv.setChartView(mChart); // For bounds control
-        mChart.setMarker(mv); // Set the marker to the chart
+        mv.setChartView(chart); // For bounds control
+        chart.setMarker(mv); // Set the marker to the chart
 
         setData();
 
-        mChart.animateXY(1400, 1400, Easing.EaseInOutQuad);
+        chart.animateXY(1400, 1400, Easing.EaseInOutQuad);
 
-        XAxis xAxis = mChart.getXAxis();
-        xAxis.setTypeface(mTfLight);
+        XAxis xAxis = chart.getXAxis();
+        xAxis.setTypeface(tfLight);
         xAxis.setTextSize(9f);
         xAxis.setYOffset(0f);
         xAxis.setXOffset(0f);
         xAxis.setValueFormatter(new IAxisValueFormatter() {
 
-            private String[] mActivities = new String[]{"Burger", "Steak", "Salad", "Pasta", "Pizza"};
+            private final String[] mActivities = new String[]{"Burger", "Steak", "Salad", "Pasta", "Pizza"};
 
             @Override
             public String getFormattedValue(float value, AxisBase axis) {
@@ -80,139 +80,41 @@ public class RadarChartActivity extends DemoBase {
         });
         xAxis.setTextColor(Color.WHITE);
 
-        YAxis yAxis = mChart.getYAxis();
-        yAxis.setTypeface(mTfLight);
+        YAxis yAxis = chart.getYAxis();
+        yAxis.setTypeface(tfLight);
         yAxis.setLabelCount(5, false);
         yAxis.setTextSize(9f);
         yAxis.setAxisMinimum(0f);
         yAxis.setAxisMaximum(80f);
         yAxis.setDrawLabels(false);
 
-        Legend l = mChart.getLegend();
+        Legend l = chart.getLegend();
         l.setVerticalAlignment(Legend.LegendVerticalAlignment.TOP);
         l.setHorizontalAlignment(Legend.LegendHorizontalAlignment.CENTER);
         l.setOrientation(Legend.LegendOrientation.HORIZONTAL);
         l.setDrawInside(false);
-        l.setTypeface(mTfLight);
+        l.setTypeface(tfLight);
         l.setXEntrySpace(7f);
         l.setYEntrySpace(5f);
         l.setTextColor(Color.WHITE);
     }
 
-    @Override
-    public boolean onCreateOptionsMenu(Menu menu) {
-        getMenuInflater().inflate(R.menu.radar, menu);
-        return true;
-    }
+    private void setData() {
 
-    @Override
-    public boolean onOptionsItemSelected(MenuItem item) {
-
-        switch (item.getItemId()) {
-            case R.id.actionToggleValues: {
-                for (IDataSet<?> set : mChart.getData().getDataSets())
-                    set.setDrawValues(!set.isDrawValuesEnabled());
-
-                mChart.invalidate();
-                break;
-            }
-            case R.id.actionToggleHighlight: {
-                if (mChart.getData() != null) {
-                    mChart.getData().setHighlightEnabled(!mChart.getData().isHighlightEnabled());
-                    mChart.invalidate();
-                }
-                break;
-            }
-            case R.id.actionToggleRotate: {
-                if (mChart.isRotationEnabled())
-                    mChart.setRotationEnabled(false);
-                else
-                    mChart.setRotationEnabled(true);
-                mChart.invalidate();
-                break;
-            }
-            case R.id.actionToggleFilled: {
-
-                ArrayList<IRadarDataSet> sets = (ArrayList<IRadarDataSet>) mChart.getData()
-                        .getDataSets();
-
-                for (IRadarDataSet set : sets) {
-                    if (set.isDrawFilledEnabled())
-                        set.setDrawFilled(false);
-                    else
-                        set.setDrawFilled(true);
-                }
-                mChart.invalidate();
-                break;
-            }
-            case R.id.actionToggleHighlightCircle: {
-
-                ArrayList<IRadarDataSet> sets = (ArrayList<IRadarDataSet>) mChart.getData()
-                        .getDataSets();
-
-                for (IRadarDataSet set : sets) {
-                    set.setDrawHighlightCircleEnabled(!set.isDrawHighlightCircleEnabled());
-                }
-                mChart.invalidate();
-                break;
-            }
-            case R.id.actionSave: {
-                if (mChart.saveToPath("title" + System.currentTimeMillis(), "")) {
-                    Toast.makeText(getApplicationContext(), "Saving SUCCESSFUL!",
-                            Toast.LENGTH_SHORT).show();
-                } else
-                    Toast.makeText(getApplicationContext(), "Saving FAILED!", Toast.LENGTH_SHORT)
-                            .show();
-                break;
-            }
-            case R.id.actionToggleXLabels: {
-                mChart.getXAxis().setEnabled(!mChart.getXAxis().isEnabled());
-                mChart.notifyDataSetChanged();
-                mChart.invalidate();
-                break;
-            }
-            case R.id.actionToggleYLabels: {
-
-                mChart.getYAxis().setEnabled(!mChart.getYAxis().isEnabled());
-                mChart.invalidate();
-                break;
-            }
-            case R.id.animateX: {
-                mChart.animateX(1400);
-                break;
-            }
-            case R.id.animateY: {
-                mChart.animateY(1400);
-                break;
-            }
-            case R.id.animateXY: {
-                mChart.animateXY(1400, 1400);
-                break;
-            }
-            case R.id.actionToggleSpin: {
-                mChart.spin(2000, mChart.getRotationAngle(), mChart.getRotationAngle() + 360, Easing.EaseInCubic);
-                break;
-            }
-        }
-        return true;
-    }
-
-    public void setData() {
-
-        float mult = 80;
+        float mul = 80;
         float min = 20;
         int cnt = 5;
 
-        ArrayList<RadarEntry> entries1 = new ArrayList<RadarEntry>();
-        ArrayList<RadarEntry> entries2 = new ArrayList<RadarEntry>();
+        ArrayList<RadarEntry> entries1 = new ArrayList<>();
+        ArrayList<RadarEntry> entries2 = new ArrayList<>();
 
         // NOTE: The order of the entries when being added to the entries array determines their position around the center of
         // the chart.
         for (int i = 0; i < cnt; i++) {
-            float val1 = (float) (Math.random() * mult) + min;
+            float val1 = (float) (Math.random() * mul) + min;
             entries1.add(new RadarEntry(val1));
 
-            float val2 = (float) (Math.random() * mult) + min;
+            float val2 = (float) (Math.random() * mul) + min;
             entries2.add(new RadarEntry(val2));
         }
 
@@ -234,17 +136,125 @@ public class RadarChartActivity extends DemoBase {
         set2.setDrawHighlightCircleEnabled(true);
         set2.setDrawHighlightIndicators(false);
 
-        ArrayList<IRadarDataSet> sets = new ArrayList<IRadarDataSet>();
+        ArrayList<IRadarDataSet> sets = new ArrayList<>();
         sets.add(set1);
         sets.add(set2);
 
         RadarData data = new RadarData(sets);
-        data.setValueTypeface(mTfLight);
+        data.setValueTypeface(tfLight);
         data.setValueTextSize(8f);
         data.setDrawValues(false);
         data.setValueTextColor(Color.WHITE);
 
-        mChart.setData(data);
-        mChart.invalidate();
+        chart.setData(data);
+        chart.invalidate();
+    }
+
+    @Override
+    public boolean onCreateOptionsMenu(Menu menu) {
+        getMenuInflater().inflate(R.menu.radar, menu);
+        return true;
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+
+        switch (item.getItemId()) {
+            case R.id.viewGithub: {
+                Intent i = new Intent(Intent.ACTION_VIEW);
+                i.setData(Uri.parse("https://github.com/PhilJay/MPAndroidChart/blob/master/MPChartExample/src/com/xxmassdeveloper/mpchartexample/RadarChartActivity.java"));
+                startActivity(i);
+                break;
+            }
+            case R.id.actionToggleValues: {
+                for (IDataSet<?> set : chart.getData().getDataSets())
+                    set.setDrawValues(!set.isDrawValuesEnabled());
+
+                chart.invalidate();
+                break;
+            }
+            case R.id.actionToggleHighlight: {
+                if (chart.getData() != null) {
+                    chart.getData().setHighlightEnabled(!chart.getData().isHighlightEnabled());
+                    chart.invalidate();
+                }
+                break;
+            }
+            case R.id.actionToggleRotate: {
+                if (chart.isRotationEnabled())
+                    chart.setRotationEnabled(false);
+                else
+                    chart.setRotationEnabled(true);
+                chart.invalidate();
+                break;
+            }
+            case R.id.actionToggleFilled: {
+
+                ArrayList<IRadarDataSet> sets = (ArrayList<IRadarDataSet>) chart.getData()
+                        .getDataSets();
+
+                for (IRadarDataSet set : sets) {
+                    if (set.isDrawFilledEnabled())
+                        set.setDrawFilled(false);
+                    else
+                        set.setDrawFilled(true);
+                }
+                chart.invalidate();
+                break;
+            }
+            case R.id.actionToggleHighlightCircle: {
+
+                ArrayList<IRadarDataSet> sets = (ArrayList<IRadarDataSet>) chart.getData()
+                        .getDataSets();
+
+                for (IRadarDataSet set : sets) {
+                    set.setDrawHighlightCircleEnabled(!set.isDrawHighlightCircleEnabled());
+                }
+                chart.invalidate();
+                break;
+            }
+            case R.id.actionToggleXLabels: {
+                chart.getXAxis().setEnabled(!chart.getXAxis().isEnabled());
+                chart.notifyDataSetChanged();
+                chart.invalidate();
+                break;
+            }
+            case R.id.actionToggleYLabels: {
+
+                chart.getYAxis().setEnabled(!chart.getYAxis().isEnabled());
+                chart.invalidate();
+                break;
+            }
+            case R.id.animateX: {
+                chart.animateX(1400);
+                break;
+            }
+            case R.id.animateY: {
+                chart.animateY(1400);
+                break;
+            }
+            case R.id.animateXY: {
+                chart.animateXY(1400, 1400);
+                break;
+            }
+            case R.id.actionToggleSpin: {
+                chart.spin(2000, chart.getRotationAngle(), chart.getRotationAngle() + 360, Easing.EaseInOutCubic);
+                break;
+            }
+            case R.id.actionSave: {
+                if (ContextCompat.checkSelfPermission(this, Manifest.permission.WRITE_EXTERNAL_STORAGE) == PackageManager.PERMISSION_GRANTED) {
+                    saveToGallery();
+                } else {
+                    requestStoragePermission(chart);
+                }
+                break;
+            }
+        }
+        return true;
+    }
+
+    @Override
+    public void saveToGallery() {
+        saveToGallery(chart, "RadarChartActivity");
     }
 }
