@@ -1,7 +1,12 @@
 
 package com.xxmassdeveloper.mpchartexample;
 
+import android.Manifest;
+import android.content.Intent;
+import android.content.pm.PackageManager;
+import android.net.Uri;
 import android.os.Bundle;
+import androidx.core.content.ContextCompat;
 import android.util.Log;
 import android.view.Menu;
 import android.view.MenuItem;
@@ -12,7 +17,6 @@ import android.widget.TextView;
 
 import com.github.mikephil.charting.charts.ScatterChart;
 import com.github.mikephil.charting.components.Legend;
-import com.github.mikephil.charting.components.Legend.LegendPosition;
 import com.github.mikephil.charting.components.XAxis;
 import com.github.mikephil.charting.components.YAxis;
 import com.github.mikephil.charting.data.Entry;
@@ -31,10 +35,10 @@ import java.util.List;
 public class ScatterChartActivity extends DemoBase implements OnSeekBarChangeListener,
         OnChartValueSelectedListener {
 
-    private ScatterChart mChart;
-    private SeekBar mSeekBarX, mSeekBarY;
+    private ScatterChart chart;
+    private SeekBar seekBarX, seekBarY;
     private TextView tvX, tvY;
-    
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -42,50 +46,107 @@ public class ScatterChartActivity extends DemoBase implements OnSeekBarChangeLis
                 WindowManager.LayoutParams.FLAG_FULLSCREEN);
         setContentView(R.layout.activity_scatterchart);
 
+        setTitle("ScatterChartActivity");
+
         tvX = findViewById(R.id.tvXMax);
         tvY = findViewById(R.id.tvYMax);
 
-        mSeekBarX = findViewById(R.id.seekBar1);
-        mSeekBarX.setOnSeekBarChangeListener(this);
+        seekBarX = findViewById(R.id.seekBar1);
+        seekBarX.setOnSeekBarChangeListener(this);
 
-        mSeekBarY = findViewById(R.id.seekBar2);
-        mSeekBarY.setOnSeekBarChangeListener(this);
+        seekBarY = findViewById(R.id.seekBar2);
+        seekBarY.setOnSeekBarChangeListener(this);
 
-        mChart = findViewById(R.id.chart1);
-        mChart.getDescription().setEnabled(false);
-        mChart.setOnChartValueSelectedListener(this);
+        chart = findViewById(R.id.chart1);
+        chart.getDescription().setEnabled(false);
+        chart.setOnChartValueSelectedListener(this);
 
-        mChart.setDrawGridBackground(false);
-        mChart.setTouchEnabled(true);
-        mChart.setMaxHighlightDistance(50f);
+        chart.setDrawGridBackground(false);
+        chart.setTouchEnabled(true);
+        chart.setMaxHighlightDistance(50f);
 
         // enable scaling and dragging
-        mChart.setDragEnabled(true);
-        mChart.setScaleEnabled(true);
+        chart.setDragEnabled(true);
+        chart.setScaleEnabled(true);
 
-        mChart.setMaxVisibleValueCount(200);
-        mChart.setPinchZoom(true);
+        chart.setMaxVisibleValueCount(200);
+        chart.setPinchZoom(true);
 
-        mSeekBarX.setProgress(45);
-        mSeekBarY.setProgress(100);
+        seekBarX.setProgress(45);
+        seekBarY.setProgress(100);
 
-        Legend l = mChart.getLegend();
+        Legend l = chart.getLegend();
         l.setVerticalAlignment(Legend.LegendVerticalAlignment.TOP);
         l.setHorizontalAlignment(Legend.LegendHorizontalAlignment.RIGHT);
         l.setOrientation(Legend.LegendOrientation.VERTICAL);
         l.setDrawInside(false);
-        l.setTypeface(mTfLight);
+        l.setTypeface(tfLight);
         l.setXOffset(5f);
 
-        YAxis yl = mChart.getAxisLeft();
-        yl.setTypeface(mTfLight);
+        YAxis yl = chart.getAxisLeft();
+        yl.setTypeface(tfLight);
         yl.setAxisMinimum(0f); // this replaces setStartAtZero(true)
-        
-        mChart.getAxisRight().setEnabled(false);
 
-        XAxis xl = mChart.getXAxis();
-        xl.setTypeface(mTfLight);
+        chart.getAxisRight().setEnabled(false);
+
+        XAxis xl = chart.getXAxis();
+        xl.setTypeface(tfLight);
         xl.setDrawGridLines(false);
+    }
+
+    @Override
+    public void onProgressChanged(SeekBar seekBar, int progress, boolean fromUser) {
+
+        tvX.setText(String.valueOf(seekBarX.getProgress()));
+        tvY.setText(String.valueOf(seekBarY.getProgress()));
+
+        ArrayList<Entry> values1 = new ArrayList<>();
+        ArrayList<Entry> values2 = new ArrayList<>();
+        ArrayList<Entry> values3 = new ArrayList<>();
+
+        for (int i = 0; i < seekBarX.getProgress(); i++) {
+            float val = (float) (Math.random() * seekBarY.getProgress()) + 3;
+            values1.add(new Entry(i, val));
+        }
+
+        for (int i = 0; i < seekBarX.getProgress(); i++) {
+            float val = (float) (Math.random() * seekBarY.getProgress()) + 3;
+            values2.add(new Entry(i+0.33f, val));
+        }
+
+        for (int i = 0; i < seekBarX.getProgress(); i++) {
+            float val = (float) (Math.random() * seekBarY.getProgress()) + 3;
+            values3.add(new Entry(i+0.66f, val));
+        }
+
+        // create a dataset and give it a type
+        ScatterDataSet set1 = new ScatterDataSet(values1, "DS 1");
+        set1.setScatterShape(ScatterChart.ScatterShape.SQUARE);
+        set1.setColor(ColorTemplate.COLORFUL_COLORS[0]);
+        ScatterDataSet set2 = new ScatterDataSet(values2, "DS 2");
+        set2.setScatterShape(ScatterChart.ScatterShape.CIRCLE);
+        set2.setScatterShapeHoleColor(ColorTemplate.COLORFUL_COLORS[3]);
+        set2.setScatterShapeHoleRadius(3f);
+        set2.setColor(ColorTemplate.COLORFUL_COLORS[1]);
+        ScatterDataSet set3 = new ScatterDataSet(values3, "DS 3");
+        set3.setShapeRenderer(new CustomScatterShapeRenderer());
+        set3.setColor(ColorTemplate.COLORFUL_COLORS[2]);
+
+        set1.setScatterShapeSize(8f);
+        set2.setScatterShapeSize(8f);
+        set3.setScatterShapeSize(8f);
+
+        ArrayList<IScatterDataSet> dataSets = new ArrayList<>();
+        dataSets.add(set1); // add the data sets
+        dataSets.add(set2);
+        dataSets.add(set3);
+
+        // create a data object with the data sets
+        ScatterData data = new ScatterData(dataSets);
+        data.setValueTypeface(tfLight);
+
+        chart.setData(data);
+        chart.invalidate();
     }
 
     @Override
@@ -98,8 +159,14 @@ public class ScatterChartActivity extends DemoBase implements OnSeekBarChangeLis
     public boolean onOptionsItemSelected(MenuItem item) {
 
         switch (item.getItemId()) {
+            case R.id.viewGithub: {
+                Intent i = new Intent(Intent.ACTION_VIEW);
+                i.setData(Uri.parse("https://github.com/PhilJay/MPAndroidChart/blob/master/MPChartExample/src/com/xxmassdeveloper/mpchartexample/ScatterChartActivity.java"));
+                startActivity(i);
+                break;
+            }
             case R.id.actionToggleValues: {
-                List<IScatterDataSet> sets = mChart.getData()
+                List<IScatterDataSet> sets = chart.getData()
                         .getDataSets();
 
                 for (IScatterDataSet iSet : sets) {
@@ -108,46 +175,49 @@ public class ScatterChartActivity extends DemoBase implements OnSeekBarChangeLis
                     set.setDrawValues(!set.isDrawValuesEnabled());
                 }
 
-                mChart.invalidate();
+                chart.invalidate();
                 break;
             }
             case R.id.actionToggleHighlight: {
-                if(mChart.getData() != null) {
-                    mChart.getData().setHighlightEnabled(!mChart.getData().isHighlightEnabled());
-                    mChart.invalidate();
+                if(chart.getData() != null) {
+                    chart.getData().setHighlightEnabled(!chart.getData().isHighlightEnabled());
+                    chart.invalidate();
                 }
                 break;
             }
             case R.id.actionTogglePinch: {
-                if (mChart.isPinchZoomEnabled())
-                    mChart.setPinchZoom(false);
+                if (chart.isPinchZoomEnabled())
+                    chart.setPinchZoom(false);
                 else
-                    mChart.setPinchZoom(true);
+                    chart.setPinchZoom(true);
 
-                mChart.invalidate();
+                chart.invalidate();
                 break;
             }
             case R.id.actionToggleAutoScaleMinMax: {
-                mChart.setAutoScaleMinMaxEnabled(!mChart.isAutoScaleMinMaxEnabled());
-                mChart.notifyDataSetChanged();
-                break;
-            }
-            case R.id.actionSave: {
-                // mChart.saveToGallery("title"+System.currentTimeMillis());
-                mChart.saveToPath("title" + System.currentTimeMillis(), "");
+                chart.setAutoScaleMinMaxEnabled(!chart.isAutoScaleMinMaxEnabled());
+                chart.notifyDataSetChanged();
                 break;
             }
             case R.id.animateX: {
-                mChart.animateX(3000);
+                chart.animateX(3000);
                 break;
             }
             case R.id.animateY: {
-                mChart.animateY(3000);
+                chart.animateY(3000);
                 break;
             }
             case R.id.animateXY: {
 
-                mChart.animateXY(3000, 3000);
+                chart.animateXY(3000, 3000);
+                break;
+            }
+            case R.id.actionSave: {
+                if (ContextCompat.checkSelfPermission(this, Manifest.permission.WRITE_EXTERNAL_STORAGE) == PackageManager.PERMISSION_GRANTED) {
+                    saveToGallery();
+                } else {
+                    requestStoragePermission(chart);
+                }
                 break;
             }
         }
@@ -155,58 +225,8 @@ public class ScatterChartActivity extends DemoBase implements OnSeekBarChangeLis
     }
 
     @Override
-    public void onProgressChanged(SeekBar seekBar, int progress, boolean fromUser) {
-
-        tvX.setText("" + (mSeekBarX.getProgress() + 1));
-        tvY.setText("" + (mSeekBarY.getProgress()));
-
-        ArrayList<Entry> yVals1 = new ArrayList<Entry>();
-        ArrayList<Entry> yVals2 = new ArrayList<Entry>();
-        ArrayList<Entry> yVals3 = new ArrayList<Entry>();
-
-        for (int i = 0; i < mSeekBarX.getProgress(); i++) {
-            float val = (float) (Math.random() * mSeekBarY.getProgress()) + 3;
-            yVals1.add(new Entry(i, val));
-        }
-
-        for (int i = 0; i < mSeekBarX.getProgress(); i++) {
-            float val = (float) (Math.random() * mSeekBarY.getProgress()) + 3;
-            yVals2.add(new Entry(i+0.33f, val));
-        }
-
-        for (int i = 0; i < mSeekBarX.getProgress(); i++) {
-            float val = (float) (Math.random() * mSeekBarY.getProgress()) + 3;
-            yVals3.add(new Entry(i+0.66f, val));
-        }
-
-        // create a dataset and give it a type
-        ScatterDataSet set1 = new ScatterDataSet(yVals1, "DS 1");
-        set1.setScatterShape(ScatterChart.ScatterShape.SQUARE);
-        set1.setColor(ColorTemplate.COLORFUL_COLORS[0]);
-        ScatterDataSet set2 = new ScatterDataSet(yVals2, "DS 2");
-        set2.setScatterShape(ScatterChart.ScatterShape.CIRCLE);
-        set2.setScatterShapeHoleColor(ColorTemplate.COLORFUL_COLORS[3]);
-        set2.setScatterShapeHoleRadius(3f);
-        set2.setColor(ColorTemplate.COLORFUL_COLORS[1]);
-        ScatterDataSet set3 = new ScatterDataSet(yVals3, "DS 3");
-        set3.setShapeRenderer(new CustomScatterShapeRenderer());
-        set3.setColor(ColorTemplate.COLORFUL_COLORS[2]);
-
-        set1.setScatterShapeSize(8f);
-        set2.setScatterShapeSize(8f);
-        set3.setScatterShapeSize(8f);
-
-        ArrayList<IScatterDataSet> dataSets = new ArrayList<IScatterDataSet>();
-        dataSets.add(set1); // add the datasets
-        dataSets.add(set2);
-        dataSets.add(set3);
-
-        // create a data object with the datasets
-        ScatterData data = new ScatterData(dataSets);
-        data.setValueTypeface(mTfLight);
-
-        mChart.setData(data);
-        mChart.invalidate();
+    public void saveToGallery() {
+        saveToGallery(chart, "ScatterChartActivity");
     }
 
     @Override
@@ -217,20 +237,11 @@ public class ScatterChartActivity extends DemoBase implements OnSeekBarChangeLis
     }
 
     @Override
-    public void onNothingSelected() {
-        // TODO Auto-generated method stub
-
-    }
+    public void onNothingSelected() {}
 
     @Override
-    public void onStartTrackingTouch(SeekBar seekBar) {
-        // TODO Auto-generated method stub
-
-    }
+    public void onStartTrackingTouch(SeekBar seekBar) {}
 
     @Override
-    public void onStopTrackingTouch(SeekBar seekBar) {
-        // TODO Auto-generated method stub
-
-    }
+    public void onStopTrackingTouch(SeekBar seekBar) {}
 }
