@@ -1,7 +1,11 @@
 
 package com.xxmassdeveloper.mpchartexample;
 
+import android.content.Intent;
+import android.net.Uri;
 import android.os.Bundle;
+import android.view.Menu;
+import android.view.MenuItem;
 import android.view.WindowManager;
 
 import com.github.mikephil.charting.charts.BarChart;
@@ -15,9 +19,10 @@ import com.xxmassdeveloper.mpchartexample.notimportant.DemoBase;
 
 import java.util.ArrayList;
 
+@SuppressWarnings("SameParameterValue")
 public class ScrollViewActivity extends DemoBase {
 
-    private BarChart mChart;
+    private BarChart chart;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -26,45 +31,71 @@ public class ScrollViewActivity extends DemoBase {
                 WindowManager.LayoutParams.FLAG_FULLSCREEN);
         setContentView(R.layout.activity_scrollview);
 
-        mChart = findViewById(R.id.chart1);
+        setTitle("ScrollViewActivity");
 
-        mChart.getDescription().setEnabled(false);
+        chart = findViewById(R.id.chart1);
+
+        chart.getDescription().setEnabled(false);
 
         // scaling can now only be done on x- and y-axis separately
-        mChart.setPinchZoom(false);
+        chart.setPinchZoom(false);
 
-        mChart.setDrawBarShadow(false);
-        mChart.setDrawGridBackground(false);
+        chart.setDrawBarShadow(false);
+        chart.setDrawGridBackground(false);
 
-        XAxis xAxis = mChart.getXAxis();
+        XAxis xAxis = chart.getXAxis();
         xAxis.setPosition(XAxisPosition.BOTTOM);
         xAxis.setDrawGridLines(false);
 
-        mChart.getAxisLeft().setDrawGridLines(false);
-        
-        mChart.getLegend().setEnabled(false);
+        chart.getAxisLeft().setDrawGridLines(false);
+
+        chart.getLegend().setEnabled(false);
 
         setData(10);
-        mChart.setFitBars(true);
+        chart.setFitBars(true);
     }
-    
+
     private void setData(int count) {
-        
-        ArrayList<BarEntry> yVals = new ArrayList<BarEntry>();
+
+        ArrayList<BarEntry> values = new ArrayList<>();
 
         for (int i = 0; i < count; i++) {
             float val = (float) (Math.random() * count) + 15;
-            yVals.add(new BarEntry(i, (int) val));
+            values.add(new BarEntry(i, (int) val));
         }
 
-        BarDataSet set = new BarDataSet(yVals, "Data Set");
+        BarDataSet set = new BarDataSet(values, "Data Set");
         set.setColors(ColorTemplate.VORDIPLOM_COLORS);
         set.setDrawValues(false);
 
         BarData data = new BarData(set);
 
-        mChart.setData(data);
-        mChart.invalidate();
-        mChart.animateY(800);
+        chart.setData(data);
+        chart.invalidate();
+        chart.animateY(800);
     }
+
+    @Override
+    public boolean onCreateOptionsMenu(Menu menu) {
+        getMenuInflater().inflate(R.menu.only_github, menu);
+        return true;
+    }
+
+    @Override
+    public boolean onOptionsItemSelected(MenuItem item) {
+
+        switch (item.getItemId()) {
+            case R.id.viewGithub: {
+                Intent i = new Intent(Intent.ACTION_VIEW);
+                i.setData(Uri.parse("https://github.com/PhilJay/MPAndroidChart/blob/master/MPChartExample/src/com/xxmassdeveloper/mpchartexample/ScrollViewActivity.java"));
+                startActivity(i);
+                break;
+            }
+        }
+
+        return true;
+    }
+
+    @Override
+    public void saveToGallery() { /* Intentionally left empty */ }
 }
