@@ -1,18 +1,21 @@
 
 package com.xxmassdeveloper.mpchartexample;
 
+import android.Manifest;
+import android.content.Intent;
+import android.content.pm.PackageManager;
 import android.graphics.Color;
+import android.net.Uri;
 import android.os.Bundle;
+import androidx.core.content.ContextCompat;
 import android.util.Log;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.WindowManager;
-import android.widget.Toast;
 
 import com.github.mikephil.charting.charts.HorizontalBarChart;
 import com.github.mikephil.charting.components.AxisBase;
 import com.github.mikephil.charting.components.Legend;
-import com.github.mikephil.charting.components.Legend.LegendPosition;
 import com.github.mikephil.charting.components.XAxis;
 import com.github.mikephil.charting.components.XAxis.XAxisPosition;
 import com.github.mikephil.charting.components.YAxis;
@@ -35,7 +38,7 @@ import java.util.List;
 public class StackedBarActivityNegative extends DemoBase implements
         OnChartValueSelectedListener {
 
-    private HorizontalBarChart mChart;
+    private HorizontalBarChart chart;
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
@@ -44,30 +47,30 @@ public class StackedBarActivityNegative extends DemoBase implements
                 WindowManager.LayoutParams.FLAG_FULLSCREEN);
         setContentView(R.layout.activity_age_distribution);
 
-        setTitle("Age Distribution Austria");
+        setTitle("StackedBarActivityNegative");
 
-        mChart = findViewById(R.id.chart1);
-        mChart.setOnChartValueSelectedListener(this);
-        mChart.setDrawGridBackground(false);
-        mChart.getDescription().setEnabled(false);
+        chart = findViewById(R.id.chart1);
+        chart.setOnChartValueSelectedListener(this);
+        chart.setDrawGridBackground(false);
+        chart.getDescription().setEnabled(false);
 
         // scaling can now only be done on x- and y-axis separately
-        mChart.setPinchZoom(false);
+        chart.setPinchZoom(false);
 
-        mChart.setDrawBarShadow(false);
-        mChart.setDrawValueAboveBar(true);
-        mChart.setHighlightFullBarEnabled(false);
-        
-        mChart.getAxisLeft().setEnabled(false);
-        mChart.getAxisRight().setAxisMaximum(25f);
-        mChart.getAxisRight().setAxisMinimum(-25f);
-        mChart.getAxisRight().setDrawGridLines(false);
-        mChart.getAxisRight().setDrawZeroLine(true);
-        mChart.getAxisRight().setLabelCount(7, false);
-        mChart.getAxisRight().setValueFormatter(new CustomFormatter());
-        mChart.getAxisRight().setTextSize(9f);
+        chart.setDrawBarShadow(false);
+        chart.setDrawValueAboveBar(true);
+        chart.setHighlightFullBarEnabled(false);
 
-        XAxis xAxis = mChart.getXAxis();
+        chart.getAxisLeft().setEnabled(false);
+        chart.getAxisRight().setAxisMaximum(25f);
+        chart.getAxisRight().setAxisMinimum(-25f);
+        chart.getAxisRight().setDrawGridLines(false);
+        chart.getAxisRight().setDrawZeroLine(true);
+        chart.getAxisRight().setLabelCount(7, false);
+        chart.getAxisRight().setValueFormatter(new CustomFormatter());
+        chart.getAxisRight().setTextSize(9f);
+
+        XAxis xAxis = chart.getXAxis();
         xAxis.setPosition(XAxisPosition.BOTH_SIDED);
         xAxis.setDrawGridLines(false);
         xAxis.setDrawAxisLine(false);
@@ -87,7 +90,7 @@ public class StackedBarActivityNegative extends DemoBase implements
             }
         });
 
-        Legend l = mChart.getLegend();
+        Legend l = chart.getLegend();
         l.setVerticalAlignment(Legend.LegendVerticalAlignment.BOTTOM);
         l.setHorizontalAlignment(Legend.LegendHorizontalAlignment.RIGHT);
         l.setOrientation(Legend.LegendOrientation.HORIZONTAL);
@@ -97,36 +100,34 @@ public class StackedBarActivityNegative extends DemoBase implements
         l.setXEntrySpace(6f);
 
         // IMPORTANT: When using negative values in stacked bars, always make sure the negative values are in the array first
-        ArrayList<BarEntry> yValues = new ArrayList<BarEntry>();
-        yValues.add(new BarEntry(5, new float[]{ -10, 10 }));
-        yValues.add(new BarEntry(15, new float[]{ -12, 13 }));
-        yValues.add(new BarEntry(25, new float[]{ -15, 15 }));
-        yValues.add(new BarEntry(35, new float[]{ -17, 17 }));
-        yValues.add(new BarEntry(45, new float[]{ -19, 20 }));
-        yValues.add(new BarEntry(45, new float[]{ -19, 20 }, getResources().getDrawable(R.drawable.star)));
-        yValues.add(new BarEntry(55, new float[]{ -19, 19 }));
-        yValues.add(new BarEntry(65, new float[]{ -16, 16 }));
-        yValues.add(new BarEntry(75, new float[]{ -13, 14 }));
-        yValues.add(new BarEntry(85, new float[]{ -10, 11 }));
-        yValues.add(new BarEntry(95, new float[]{ -5, 6 }));
-        yValues.add(new BarEntry(105, new float[]{ -1, 2 }));
+        ArrayList<BarEntry> values = new ArrayList<>();
+        values.add(new BarEntry(5, new float[]{ -10, 10 }));
+        values.add(new BarEntry(15, new float[]{ -12, 13 }));
+        values.add(new BarEntry(25, new float[]{ -15, 15 }));
+        values.add(new BarEntry(35, new float[]{ -17, 17 }));
+        values.add(new BarEntry(45, new float[]{ -19, 20 }));
+        values.add(new BarEntry(45, new float[]{ -19, 20 }, getResources().getDrawable(R.drawable.star)));
+        values.add(new BarEntry(55, new float[]{ -19, 19 }));
+        values.add(new BarEntry(65, new float[]{ -16, 16 }));
+        values.add(new BarEntry(75, new float[]{ -13, 14 }));
+        values.add(new BarEntry(85, new float[]{ -10, 11 }));
+        values.add(new BarEntry(95, new float[]{ -5, 6 }));
+        values.add(new BarEntry(105, new float[]{ -1, 2 }));
 
-        BarDataSet set = new BarDataSet(yValues, "Age Distribution");
+        BarDataSet set = new BarDataSet(values, "Age Distribution");
         set.setDrawIcons(false);
         set.setValueFormatter(new CustomFormatter());
         set.setValueTextSize(7f);
         set.setAxisDependency(YAxis.AxisDependency.RIGHT);
-        set.setColors(new int[] {Color.rgb(67,67,72), Color.rgb(124,181,236)});
+        set.setColors(Color.rgb(67,67,72), Color.rgb(124,181,236));
         set.setStackLabels(new String[]{
                 "Men", "Women"
         });
 
-        String []xLabels = new String[]{"0-10", "10-20", "20-30", "30-40", "40-50", "50-60", "60-70", "70-80", "80-90", "90-100", "100+"};
-
         BarData data = new BarData(set);
         data.setBarWidth(8.5f);
-        mChart.setData(data);
-        mChart.invalidate();
+        chart.setData(data);
+        chart.invalidate();
     }
 
     @Override
@@ -139,8 +140,14 @@ public class StackedBarActivityNegative extends DemoBase implements
     public boolean onOptionsItemSelected(MenuItem item) {
 
         switch (item.getItemId()) {
+            case R.id.viewGithub: {
+                Intent i = new Intent(Intent.ACTION_VIEW);
+                i.setData(Uri.parse("https://github.com/PhilJay/MPAndroidChart/blob/master/MPChartExample/src/com/xxmassdeveloper/mpchartexample/StackedBarActivityNegative.java"));
+                startActivity(i);
+                break;
+            }
             case R.id.actionToggleValues: {
-                List<IBarDataSet> sets = mChart.getData()
+                List<IBarDataSet> sets = chart.getData()
                         .getDataSets();
 
                 for (IBarDataSet iSet : sets) {
@@ -149,11 +156,11 @@ public class StackedBarActivityNegative extends DemoBase implements
                     set.setDrawValues(!set.isDrawValuesEnabled());
                 }
 
-                mChart.invalidate();
+                chart.invalidate();
                 break;
             }
             case R.id.actionToggleIcons: {
-                List<IBarDataSet> sets = mChart.getData()
+                List<IBarDataSet> sets = chart.getData()
                         .getDataSets();
 
                 for (IBarDataSet iSet : sets) {
@@ -162,57 +169,56 @@ public class StackedBarActivityNegative extends DemoBase implements
                     set.setDrawIcons(!set.isDrawIconsEnabled());
                 }
 
-                mChart.invalidate();
+                chart.invalidate();
                 break;
             }
             case R.id.actionToggleHighlight: {
-                if(mChart.getData() != null) {
-                    mChart.getData().setHighlightEnabled(!mChart.getData().isHighlightEnabled());
-                    mChart.invalidate();
+                if(chart.getData() != null) {
+                    chart.getData().setHighlightEnabled(!chart.getData().isHighlightEnabled());
+                    chart.invalidate();
                 }
                 break;
             }
             case R.id.actionTogglePinch: {
-                if (mChart.isPinchZoomEnabled())
-                    mChart.setPinchZoom(false);
+                if (chart.isPinchZoomEnabled())
+                    chart.setPinchZoom(false);
                 else
-                    mChart.setPinchZoom(true);
+                    chart.setPinchZoom(true);
 
-                mChart.invalidate();
+                chart.invalidate();
                 break;
             }
             case R.id.actionToggleAutoScaleMinMax: {
-                mChart.setAutoScaleMinMaxEnabled(!mChart.isAutoScaleMinMaxEnabled());
-                mChart.notifyDataSetChanged();
+                chart.setAutoScaleMinMaxEnabled(!chart.isAutoScaleMinMaxEnabled());
+                chart.notifyDataSetChanged();
                 break;
             }
             case R.id.actionToggleBarBorders: {
-                for (IBarDataSet set : mChart.getData().getDataSets())
+                for (IBarDataSet set : chart.getData().getDataSets())
                     ((BarDataSet)set).setBarBorderWidth(set.getBarBorderWidth() == 1.f ? 0.f : 1.f);
 
-                mChart.invalidate();
+                chart.invalidate();
                 break;
             }
             case R.id.animateX: {
-                mChart.animateX(3000);
+                chart.animateX(3000);
                 break;
             }
             case R.id.animateY: {
-                mChart.animateY(3000);
+                chart.animateY(3000);
                 break;
             }
             case R.id.animateXY: {
 
-                mChart.animateXY(3000, 3000);
+                chart.animateXY(3000, 3000);
                 break;
             }
             case R.id.actionSave: {
-                if (mChart.saveToGallery("title" + System.currentTimeMillis(), 50)) {
-                    Toast.makeText(getApplicationContext(), "Saving SUCCESSFUL!",
-                            Toast.LENGTH_SHORT).show();
-                } else
-                    Toast.makeText(getApplicationContext(), "Saving FAILED!", Toast.LENGTH_SHORT)
-                            .show();
+                if (ContextCompat.checkSelfPermission(this, Manifest.permission.WRITE_EXTERNAL_STORAGE) == PackageManager.PERMISSION_GRANTED) {
+                    saveToGallery();
+                } else {
+                    requestStoragePermission(chart);
+                }
                 break;
             }
         }
@@ -220,8 +226,12 @@ public class StackedBarActivityNegative extends DemoBase implements
     }
 
     @Override
-    public void onValueSelected(Entry e, Highlight h) {
+    public void saveToGallery() {
+        saveToGallery(chart, "StackedBarActivityNegative");
+    }
 
+    @Override
+    public void onValueSelected(Entry e, Highlight h) {
         BarEntry entry = (BarEntry) e;
         Log.i("VAL SELECTED",
                 "Value: " + Math.abs(entry.getYVals()[h.getStackIndex()]));
@@ -229,16 +239,14 @@ public class StackedBarActivityNegative extends DemoBase implements
 
     @Override
     public void onNothingSelected() {
-        // TODO Auto-generated method stub
         Log.i("NOTING SELECTED", "");
     }
 
-    private class CustomFormatter implements IValueFormatter, IAxisValueFormatter
-    {
+    private class CustomFormatter implements IValueFormatter, IAxisValueFormatter {
 
         private DecimalFormat mFormat;
 
-        public CustomFormatter() {
+        CustomFormatter() {
             mFormat = new DecimalFormat("###");
         }
 
