@@ -1,6 +1,7 @@
 
 package com.xxmassdeveloper.mpchartexample.custom;
 
+import android.annotation.SuppressLint;
 import android.content.Context;
 import android.widget.TextView;
 
@@ -14,9 +15,11 @@ import com.xxmassdeveloper.mpchartexample.R;
 
 /**
  * Custom implementation of the MarkerView.
- * 
+ *
  * @author Philipp Jahoda
  */
+@SuppressWarnings("unused")
+@SuppressLint("ViewConstructor")
 public class StackedBarsMarkerView extends MarkerView {
 
     private TextView tvContent;
@@ -27,7 +30,7 @@ public class StackedBarsMarkerView extends MarkerView {
         tvContent = findViewById(R.id.tvContent);
     }
 
-    // callbacks everytime the MarkerView is redrawn, can be used to update the
+    // runs every time the MarkerView is redrawn, can be used to update the
     // content (user-interface)
     @Override
     public void refreshContent(Entry e, Highlight highlight) {
@@ -39,13 +42,13 @@ public class StackedBarsMarkerView extends MarkerView {
             if(be.getYVals() != null) {
 
                 // draw the stack value
-                tvContent.setText("" + Utils.formatNumber(be.getYVals()[highlight.getStackIndex()], 0, true));
+                tvContent.setText(Utils.formatNumber(be.getYVals()[highlight.getStackIndex()], 0, true));
             } else {
-                tvContent.setText("" + Utils.formatNumber(be.getY(), 0, true));
+                tvContent.setText(Utils.formatNumber(be.getY(), 0, true));
             }
         } else {
 
-            tvContent.setText("" + Utils.formatNumber(e.getY(), 0, true));
+            tvContent.setText(Utils.formatNumber(e.getY(), 0, true));
         }
 
         super.refreshContent(e, highlight);
