@@ -1,5 +1,6 @@
 package com.xxmassdeveloper.mpchartexample.listviewitems;
 
+import android.annotation.SuppressLint;
 import android.content.Context;
 import android.graphics.Typeface;
 import android.view.LayoutInflater;
@@ -14,9 +15,9 @@ import com.github.mikephil.charting.data.ChartData;
 import com.xxmassdeveloper.mpchartexample.R;
 
 public class BarChartItem extends ChartItem {
-    
+
     private Typeface mTf;
-    
+
     public BarChartItem(ChartData<?> cd, Context c) {
         super(cd);
 
@@ -28,10 +29,11 @@ public class BarChartItem extends ChartItem {
         return TYPE_BARCHART;
     }
 
+    @SuppressLint("InflateParams")
     @Override
     public View getView(int position, View convertView, Context c) {
 
-        ViewHolder holder = null;
+        ViewHolder holder;
 
         if (convertView == null) {
 
@@ -57,13 +59,13 @@ public class BarChartItem extends ChartItem {
         xAxis.setTypeface(mTf);
         xAxis.setDrawGridLines(false);
         xAxis.setDrawAxisLine(true);
-        
+
         YAxis leftAxis = holder.chart.getAxisLeft();
         leftAxis.setTypeface(mTf);
         leftAxis.setLabelCount(5, false);
         leftAxis.setSpaceTop(20f);
         leftAxis.setAxisMinimum(0f); // this replaces setStartAtZero(true)
-       
+
         YAxis rightAxis = holder.chart.getAxisRight();
         rightAxis.setTypeface(mTf);
         rightAxis.setLabelCount(5, false);
@@ -71,18 +73,18 @@ public class BarChartItem extends ChartItem {
         rightAxis.setAxisMinimum(0f); // this replaces setStartAtZero(true)
 
         mChartData.setValueTypeface(mTf);
-        
+
         // set data
         holder.chart.setData((BarData) mChartData);
         holder.chart.setFitBars(true);
-        
+
         // do not forget to refresh the chart
 //        holder.chart.invalidate();
         holder.chart.animateY(700);
 
         return convertView;
     }
-    
+
     private static class ViewHolder {
         BarChart chart;
     }
