@@ -6,23 +6,24 @@ import android.view.View;
 import com.github.mikephil.charting.data.ChartData;
 
 /**
- * baseclass of the chart-listview items
+ * Base class of the Chart ListView items
  * @author philipp
  *
  */
+@SuppressWarnings("unused")
 public abstract class ChartItem {
-    
-    protected static final int TYPE_BARCHART = 0;
-    protected static final int TYPE_LINECHART = 1;
-    protected static final int TYPE_PIECHART = 2;
-    
-    protected ChartData<?> mChartData;
-    
-    public ChartItem(ChartData<?> cd) {
-        this.mChartData = cd;      
+
+    static final int TYPE_BARCHART = 0;
+    static final int TYPE_LINECHART = 1;
+    static final int TYPE_PIECHART = 2;
+
+    ChartData<?> mChartData;
+
+    ChartItem(ChartData<?> cd) {
+        this.mChartData = cd;
     }
-    
+
     public abstract int getItemType();
-    
+
     public abstract View getView(int position, View convertView, Context c);
 }
